Get record from the Pangaea repository

Get record from the Pangaea repository

pg_get_record(identifier, prefix = "oai_dc", as = "df", ...)

Arguments

identifier

Dataset identifier. See Examples.

prefix

A character string to specify the metadata format in OAI-PMH requests issued to the repository. The default (oai_dc) corresponds to the mandatory OAI unqualified Dublin Core metadata schema.

as

(character) What to return. One of "df" (for data.frame; default), "list", or "raw" (raw text)

...

Curl debugging options passed on to httr::GET()

Value

XML character string, data.frame, or list, depending on what requested with the as parameter

References

https://www.openarchives.org/pmh/

See also

wraps oai::get_records()

Other oai methods: pg_identify, pg_list_identifiers, pg_list_metadata_formats, pg_list_records, pg_list_sets

Examples

pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.788382")

#> $`oai:pangaea.de:doi:10.1594/PANGAEA.788382`
#> $`oai:pangaea.de:doi:10.1594/PANGAEA.788382`$header
#> # A tibble: 1 × 3
#>                                  identifier            datestamp
#>                                       <chr>                <chr>
#> 1 oai:pangaea.de:doi:10.1594/PANGAEA.788382 2016-06-25T14:58:45Z
#> # ... with 1 more variables: setSpec <chr>
#> 
#> $`oai:pangaea.de:doi:10.1594/PANGAEA.788382`$metadata
#> # A tibble: 1 × 13
#>                                                                              title
#>                                                                              <chr>
#> 1 Trace metals in shells of mussels and clams from deep-sea hydrothermal vent fiel
#> # ... with 12 more variables: creator <chr>, source <chr>, publisher <chr>,
#> #   date <chr>, type <chr>, format <chr>, identifier <chr>, description <chr>,
#> #   language <chr>, rights <chr>, coverage <chr>, subject <chr>
#> 
#> 

pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.269656",
prefix="iso19139")

#> $`oai:pangaea.de:doi:10.1594/PANGAEA.269656`
#> [1] "<?xml version=\"1.0\" encoding=\"UTF-8\"?><?xml-stylesheet type=\"text/xsl\" href=\"oai2.xsl\"?><OAI-PMH xmlns=\"http://www.openarchives.org/OAI/2.0/\" xsi:schemaLocation=\"http://www.openarchives.org/OAI/2.0/ http://www.openarchives.org/OAI/2.0/OAI-PMH.xsd\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\"><responseDate>2017-03-09T19:10:42Z</responseDate><request verb=\"GetRecord\" metadataPrefix=\"iso19139\" identifier=\"oai:pangaea.de:doi:10.1594/PANGAEA.269656\">https://ws.pangaea.de/oai/provider</request><GetRecord><record><header><identifier>oai:pangaea.de:doi:10.1594/PANGAEA.269656</identifier><datestamp>2016-12-16T22:01:16Z</datestamp></header><metadata><MD_Metadata xmlns:xlink=\"http://www.w3.org/1999/xlink\" xmlns:gmd=\"http://www.isotc211.org/2005/gmd\" xmlns:gco=\"http://www.isotc211.org/2005/gco\" xmlns:gml=\"http://www.opengis.net/gml\" xmlns=\"http://www.isotc211.org/2005/gmd\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xsi:schemaLocation=\"http://www.isotc211.org/2005/gmd http://www.isotc211.org/2005/gmd/gmd.xsd\" id=\"de.pangaea.dataset269656\"><fileIdentifier><gco:CharacterString>de.pangaea.dataset269656</gco:CharacterString></fileIdentifier><contact><CI_ResponsibleParty><organisationName><gco:CharacterString>PANGAEA</gco:CharacterString></organisationName><contactInfo><CI_Contact><address><CI_Address><electronicMailAddress><gco:CharacterString>info@pangaea.de</gco:CharacterString></electronicMailAddress></CI_Address></address><onlineResource><CI_OnlineResource><linkage><URL>http://www.pangaea.de/</URL></linkage><function><CI_OnLineFunctionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode_information\">information</CI_OnLineFunctionCode></function></CI_OnlineResource></onlineResource></CI_Contact></contactInfo><role><CI_RoleCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode_pointOfContact\">pointOfContact</CI_RoleCode></role></CI_ResponsibleParty></contact><dateStamp><gco:DateTime>2016-12-16T22:01:16Z</gco:DateTime></dateStamp><dataSetURI><gco:CharacterString>https://doi.pangaea.de/10.1594/PANGAEA.269656</gco:CharacterString></dataSetURI><identificationInfo><MD_DataIdentification><citation><CI_Citation><title><gco:CharacterString>Meteorological synoptical observations from Neumayer Station (1993-12)</gco:CharacterString></title><date><CI_Date><date><gco:DateTime>2005-05-10T09:15:33</gco:DateTime></date><dateType><CI_DateTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_DateTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_DateTypeCode_publication\">publication</CI_DateTypeCode></dateType></CI_Date></date><identifier><MD_Identifier><code><gco:CharacterString>doi:10.1594/PANGAEA.269656</gco:CharacterString></code></MD_Identifier></identifier><citedResponsibleParty><CI_ResponsibleParty id=\"dataset.author10396\"><individualName><gco:CharacterString>König-Langlo, Gert</gco:CharacterString></individualName><contactInfo><CI_Contact><address><CI_Address><electronicMailAddress><gco:CharacterString>gert.koenig-langlo@awi.de</gco:CharacterString></electronicMailAddress></CI_Address></address><onlineResource><CI_OnlineResource><linkage><URL>http://www.awi.de/en/about-us/organisation/staff/gert-koenig-langlo.html</URL></linkage><function><CI_OnLineFunctionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode_information\">information</CI_OnLineFunctionCode></function></CI_OnlineResource></onlineResource></CI_Contact></contactInfo><role><CI_RoleCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode_author\">author</CI_RoleCode></role></CI_ResponsibleParty></citedResponsibleParty><citedResponsibleParty><CI_ResponsibleParty><organisationName><gco:CharacterString>PANGAEA</gco:CharacterString></organisationName><contactInfo><CI_Contact><address><CI_Address><electronicMailAddress><gco:CharacterString>info@pangaea.de</gco:CharacterString></electronicMailAddress></CI_Address></address><onlineResource><CI_OnlineResource><linkage><URL>http://www.pangaea.de/</URL></linkage><function><CI_OnLineFunctionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode_information\">information</CI_OnLineFunctionCode></function></CI_OnlineResource></onlineResource></CI_Contact></contactInfo><role><CI_RoleCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode_publisher\">publisher</CI_RoleCode></role></CI_ResponsibleParty></citedResponsibleParty><citedResponsibleParty><CI_ResponsibleParty><organisationName><gco:CharacterString>Alfred Wegener Institute, Helmholtz Center for Polar and Marine Research, Bremerhaven</gco:CharacterString></organisationName><role><CI_RoleCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode_originator\">originator</CI_RoleCode></role></CI_ResponsibleParty></citedResponsibleParty><presentationForm><CI_PresentationFormCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_PresentationFormCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_PresentationFormCode_tableDigital\">tableDigital</CI_PresentationFormCode></presentationForm></CI_Citation></citation><abstract gco:nilReason=\"missing\"/><status><MD_ProgressCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_ProgressCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_ProgressCode_completed\">completed</MD_ProgressCode></status><pointOfContact><CI_ResponsibleParty id=\"pi10396\"><individualName><gco:CharacterString>König-Langlo, Gert</gco:CharacterString></individualName><contactInfo><CI_Contact><address><CI_Address><electronicMailAddress><gco:CharacterString>gert.koenig-langlo@awi.de</gco:CharacterString></electronicMailAddress></CI_Address></address><onlineResource><CI_OnlineResource><linkage><URL>http://www.awi.de/en/about-us/organisation/staff/gert-koenig-langlo.html</URL></linkage><function><CI_OnLineFunctionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode_information\">information</CI_OnLineFunctionCode></function></CI_OnlineResource></onlineResource></CI_Contact></contactInfo><role><CI_RoleCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode_principalInvestigator\">principalInvestigator</CI_RoleCode></role></CI_ResponsibleParty></pointOfContact><descriptiveKeywords><MD_Keywords><keyword><gco:CharacterString>BSRN</gco:CharacterString></keyword><keyword><gco:CharacterString>Baseline Surface Radiation Network</gco:CharacterString></keyword><type><MD_KeywordTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_KeywordTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_KeywordTypeCode_theme\">theme</MD_KeywordTypeCode></type><thesaurusName><CI_Citation><title><gco:CharacterString>PANGAEA Project List</gco:CharacterString></title><date><CI_Date><date><gco:Date>1995-01-01</gco:Date></date><dateType><CI_DateTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_DateTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_DateTypeCode_creation\">creation</CI_DateTypeCode></dateType></CI_Date></date><citedResponsibleParty><CI_ResponsibleParty><organisationName><gco:CharacterString>PANGAEA</gco:CharacterString></organisationName><contactInfo><CI_Contact><address><CI_Address><electronicMailAddress><gco:CharacterString>info@pangaea.de</gco:CharacterString></electronicMailAddress></CI_Address></address><onlineResource><CI_OnlineResource><linkage><URL>http://www.pangaea.de/</URL></linkage><function><CI_OnLineFunctionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode_information\">information</CI_OnLineFunctionCode></function></CI_OnlineResource></onlineResource></CI_Contact></contactInfo><role><CI_RoleCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode_pointOfContact\">pointOfContact</CI_RoleCode></role></CI_ResponsibleParty></citedResponsibleParty></CI_Citation></thesaurusName></MD_Keywords></descriptiveKeywords><resourceConstraints><MD_LegalConstraints><accessConstraints><MD_RestrictionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_RestrictionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_RestrictionCode_copyright\">copyright</MD_RestrictionCode></accessConstraints><accessConstraints><MD_RestrictionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_RestrictionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_RestrictionCode_restricted\">restricted</MD_RestrictionCode></accessConstraints><useConstraints><MD_RestrictionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_RestrictionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_RestrictionCode_copyright\">copyright</MD_RestrictionCode></useConstraints></MD_LegalConstraints></resourceConstraints><spatialRepresentationType><MD_SpatialRepresentationTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_SpatialRepresentationTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_SpatialRepresentationTypeCode_textTable\">textTable</MD_SpatialRepresentationTypeCode></spatialRepresentationType><language><LanguageCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/ML_gmxCodelists.xml#LanguageCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/ML_gmxCodelists.xml#LanguageCode_eng\">eng</LanguageCode></language><topicCategory><MD_TopicCategoryCode>geoscientificInformation</MD_TopicCategoryCode></topicCategory><extent><EX_Extent><geographicElement><EX_GeographicBoundingBox><gmd:westBoundLongitude><gco:Decimal>-8.25</gco:Decimal></gmd:westBoundLongitude><gmd:eastBoundLongitude><gco:Decimal>-8.25</gco:Decimal></gmd:eastBoundLongitude><gmd:southBoundLatitude><gco:Decimal>-70.65</gco:Decimal></gmd:southBoundLatitude><gmd:northBoundLatitude><gco:Decimal>-70.65</gco:Decimal></gmd:northBoundLatitude></EX_GeographicBoundingBox></geographicElement><temporalElement><EX_TemporalExtent><extent><gml:TimePeriod gml:id=\"timePeriod\"><gml:beginPosition>1993-12-01T00:00:00</gml:beginPosition><gml:endPosition>1993-12-31T21:00:00</gml:endPosition></gml:TimePeriod></extent></EX_TemporalExtent></temporalElement></EX_Extent></extent></MD_DataIdentification></identificationInfo><contentInfo><MD_CoverageDescription id=\"col0.ds1096525\"><attributeDescription><gco:RecordType>DATE/TIME</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_thematicClassification\">thematicClassification</MD_CoverageContentTypeCode></contentType></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col1.ds10240425\"><attributeDescription><gco:RecordType>Cloud base height</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col1.ds10240425.band\"><descriptor><gco:CharacterString>Ceilometer</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col1.ds10240425.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col2.ds10240426\"><attributeDescription><gco:RecordType>Horizontal visibility</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col2.ds10240426.band\"><descriptor><gco:CharacterString>Visibility sensor</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col2.ds10240426.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col3.ds10240427\"><attributeDescription><gco:RecordType>Wind direction</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col3.ds10240427.band\"><descriptor><gco:CharacterString>Anemometer</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col3.ds10240427.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">deg</gml:identifier><gml:name>deg</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col4.ds10240428\"><attributeDescription><gco:RecordType>Wind speed</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col4.ds10240428.band\"><descriptor><gco:CharacterString>Anemometer</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col4.ds10240428.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">m/s</gml:identifier><gml:name>m/s</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col5.ds10240429\"><attributeDescription><gco:RecordType>Temperature, air</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col5.ds10240429.band\"><descriptor><gco:CharacterString>Thermometer</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col5.ds10240429.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">°C</gml:identifier><gml:name>°C</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col6.ds10240430\"><attributeDescription><gco:RecordType>Dew/frost point</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col6.ds10240430.band\"><descriptor><gco:CharacterString>Hygrometer</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col6.ds10240430.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">°C</gml:identifier><gml:name>°C</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col7.ds10240431\"><attributeDescription><gco:RecordType>Pressure, atmospheric</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col7.ds10240431.band\"><descriptor><gco:CharacterString>Barometer; Station pressure reduced to sea level</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col7.ds10240431.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">hPa</gml:identifier><gml:name>hPa</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col8.ds10240432\"><attributeDescription><gco:RecordType>Characteristic of barometric tendency</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col8.ds10240432.band\"><descriptor><gco:CharacterString>Barometer</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col8.ds10240432.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col9.ds10240433\"><attributeDescription><gco:RecordType>Amount of barometric tendency</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col9.ds10240433.band\"><descriptor><gco:CharacterString>Barometer</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col9.ds10240433.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">hPa</gml:identifier><gml:name>hPa</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col10.ds10240434\"><attributeDescription><gco:RecordType>Present weather</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col10.ds10240434.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col10.ds10240434.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col11.ds10240435\"><attributeDescription><gco:RecordType>Past weather1</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col11.ds10240435.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col11.ds10240435.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col12.ds10240436\"><attributeDescription><gco:RecordType>Past weather2</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col12.ds10240436.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col12.ds10240436.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col13.ds10240437\"><attributeDescription><gco:RecordType>Low cloud</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col13.ds10240437.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col13.ds10240437.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col14.ds10240438\"><attributeDescription><gco:RecordType>Middle cloud</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col14.ds10240438.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col14.ds10240438.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col15.ds10240439\"><attributeDescription><gco:RecordType>High cloud</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col15.ds10240439.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col15.ds10240439.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col16.ds10240440\"><attributeDescription><gco:RecordType>Total cloud amount</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col16.ds10240440.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col16.ds10240440.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col17.ds10240441\"><attributeDescription><gco:RecordType>Low/middle cloud amount</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col17.ds10240441.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col17.ds10240441.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col18.ds10240442\"><attributeDescription><gco:RecordType>Temperature, air, maximum</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col18.ds10240442.band\"><descriptor><gco:CharacterString>Thermometer</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col18.ds10240442.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">°C</gml:identifier><gml:name>°C</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col19.ds10240443\"><attributeDescription><gco:RecordType>Temperature, air, minimum</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col19.ds10240443.band\"><descriptor><gco:CharacterString>Thermometer</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col19.ds10240443.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">°C</gml:identifier><gml:name>°C</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col20.ds10240444\"><attributeDescription><gco:RecordType>Present blowing snow</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col20.ds10240444.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col20.ds10240444.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col21.ds10240445\"><attributeDescription><gco:RecordType>Past blowing snow</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col21.ds10240445.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col21.ds10240445.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">code</gml:identifier><gml:name>code</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><contentInfo><MD_CoverageDescription id=\"col22.ds10240446\"><attributeDescription><gco:RecordType>Whiteout yes/no</gco:RecordType></attributeDescription><contentType><MD_CoverageContentTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_CoverageContentTypeCode_physicalMeasurement\">physicalMeasurement</MD_CoverageContentTypeCode></contentType><dimension><MD_Band id=\"col22.ds10240446.band\"><descriptor><gco:CharacterString>Visual observation</gco:CharacterString></descriptor><units><gml:UnitDefinition gml:id=\"col22.ds10240446.band.unit\"><gml:identifier codeSpace=\"PANGAEA\">y/n</gml:identifier><gml:name>y/n</gml:name></gml:UnitDefinition></units></MD_Band></dimension></MD_CoverageDescription></contentInfo><distributionInfo><MD_Distribution><distributionFormat><MD_Format><name><gco:CharacterString>text/tab-separated-values</gco:CharacterString></name><version><gco:CharacterString>unknown</gco:CharacterString></version></MD_Format></distributionFormat><transferOptions><MD_DigitalTransferOptions><onLine><CI_OnlineResource><linkage><URL>https://doi.pangaea.de/10.1594/PANGAEA.269656?format=textfile</URL></linkage><function><CI_OnLineFunctionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode_download\">download</CI_OnLineFunctionCode></function></CI_OnlineResource></onLine></MD_DigitalTransferOptions></transferOptions></MD_Distribution></distributionInfo><dataQualityInfo><DQ_DataQuality><scope><DQ_Scope><level><MD_ScopeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_ScopeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#MD_ScopeCode_dataset\">dataset</MD_ScopeCode></level></DQ_Scope></scope><lineage><LI_Lineage><statement><gco:CharacterString>The data set was checked for completeness, correctness, and consistency of metainformation. Validity of used methods was checked and - if applicable - precision and range of data.</gco:CharacterString></statement><processStep><LI_ProcessStep id=\"event2536919\"><description><gco:CharacterString>EVENT LABEL:  (Georg von Neumayer) * LATITUDE: -70.650000 * LONGITUDE: -8.250000 * DATE/TIME: 1992-01-01T00:00:00 * ELEVATION: 42.0 m * LOCATION: Dronning Maud Land, Antarctica * CAMPAIGN: WCRP/GEWEX * DEVICE: Monitoring station</gco:CharacterString></description><rationale><gco:CharacterString>BSRN station no: 13; Surface type: iceshelf; Topography type: flat, rural; Horizon from 1992 to 2009-01: doi:10.1594/PANGAEA.669516; Horizon after 2009-01: doi:10.1594/PANGAEA.757811; Station scientist: Gert König-Langlo (Gert.Koenig-Langlo@awi.de). Station description see hdl:10013/epic.28566.d001</gco:CharacterString></rationale><dateTime><gco:DateTime>1992-01-01T00:00:00</gco:DateTime></dateTime></LI_ProcessStep></processStep><source><LI_Source id=\"ref40936\"><description><gco:CharacterString>Further details</gco:CharacterString></description><sourceCitation><CI_Citation><title><gco:CharacterString>Validation routines for synoptic observations</gco:CharacterString></title><date><CI_Date><date><gco:Date>2012-01-01</gco:Date></date><dateType><CI_DateTypeCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_DateTypeCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_DateTypeCode_publication\">publication</CI_DateTypeCode></dateType></CI_Date></date><identifier><MD_Identifier><code><gco:CharacterString>hdl:10013/epic.40293.d001</gco:CharacterString></code></MD_Identifier></identifier><citedResponsibleParty><CI_ResponsibleParty id=\"ref40936.author10396\"><individualName><gco:CharacterString>König-Langlo, Gert</gco:CharacterString></individualName><contactInfo><CI_Contact><address><CI_Address><electronicMailAddress><gco:CharacterString>gert.koenig-langlo@awi.de</gco:CharacterString></electronicMailAddress></CI_Address></address><onlineResource><CI_OnlineResource><linkage><URL>http://www.awi.de/en/about-us/organisation/staff/gert-koenig-langlo.html</URL></linkage><function><CI_OnLineFunctionCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_OnLineFunctionCode_information\">information</CI_OnLineFunctionCode></function></CI_OnlineResource></onlineResource></CI_Contact></contactInfo><role><CI_RoleCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_RoleCode_author\">author</CI_RoleCode></role></CI_ResponsibleParty></citedResponsibleParty><presentationForm><CI_PresentationFormCode codeList=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_PresentationFormCode\" codeListValue=\"http://www.isotc211.org/2005/resources/Codelist/gmxCodelists.xml#CI_PresentationFormCode_documentDigital\">documentDigital</CI_PresentationFormCode></presentationForm><series><CI_Series><name><gco:CharacterString>Alfred Wegener Institute for Polar and Marine Research, Bremerhaven, Germany</gco:CharacterString></name></CI_Series></series></CI_Citation></sourceCitation></LI_Source></source></LI_Lineage></lineage></DQ_DataQuality></dataQualityInfo></MD_Metadata></metadata></record></GetRecord></OAI-PMH>"
#> 

pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.269656",
prefix="dif")

#> $`oai:pangaea.de:doi:10.1594/PANGAEA.269656`
#> [1] "<?xml version=\"1.0\" encoding=\"UTF-8\"?><?xml-stylesheet type=\"text/xsl\" href=\"oai2.xsl\"?><OAI-PMH xmlns=\"http://www.openarchives.org/OAI/2.0/\" xsi:schemaLocation=\"http://www.openarchives.org/OAI/2.0/ http://www.openarchives.org/OAI/2.0/OAI-PMH.xsd\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\"><responseDate>2017-03-09T19:10:42Z</responseDate><request verb=\"GetRecord\" metadataPrefix=\"dif\" identifier=\"oai:pangaea.de:doi:10.1594/PANGAEA.269656\">https://ws.pangaea.de/oai/provider</request><GetRecord><record><header><identifier>oai:pangaea.de:doi:10.1594/PANGAEA.269656</identifier><datestamp>2016-12-16T22:01:16Z</datestamp></header><metadata><DIF xmlns=\"http://gcmd.gsfc.nasa.gov/Aboutus/xml/dif/\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xsi:schemaLocation=\"http://gcmd.gsfc.nasa.gov/Aboutus/xml/dif/ http://gcmd.gsfc.nasa.gov/Aboutus/xml/dif/dif_v9.4.xsd\"><Entry_ID>PANGAEA_269656</Entry_ID><Entry_Title>Meteorological synoptical observations from Neumayer Station (1993-12)</Entry_Title><Data_Set_Citation><Dataset_Creator>König-Langlo, Gert</Dataset_Creator><Dataset_Title>Meteorological synoptical observations from Neumayer Station (1993-12)</Dataset_Title><Dataset_Release_Date>2005-05-10</Dataset_Release_Date><Dataset_Publisher>PANGAEA</Dataset_Publisher><Data_Presentation_Form>Dataset</Data_Presentation_Form><Online_Resource>https://doi.pangaea.de/10.1594/PANGAEA.269656</Online_Resource></Data_Set_Citation><Personnel><Role>Investigator</Role><First_Name>Gert</First_Name><Last_Name>König-Langlo</Last_Name><Email>gert.koenig-langlo@awi.de</Email></Personnel><Discipline><Discipline_Name>Earth Science</Discipline_Name></Discipline><Parameters><Detailed_Variable>DATE/TIME</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Cloud base height</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Horizontal visibility</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Wind direction</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Wind speed</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Temperature, air</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Dew/frost point</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Pressure, atmospheric</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Characteristic of barometric tendency</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Amount of barometric tendency</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Present weather</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Past weather1</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Past weather2</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Low cloud</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Middle cloud</Detailed_Variable></Parameters><Parameters><Detailed_Variable>High cloud</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Total cloud amount</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Low/middle cloud amount</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Temperature, air, maximum</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Temperature, air, minimum</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Present blowing snow</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Past blowing snow</Detailed_Variable></Parameters><Parameters><Detailed_Variable>Whiteout yes/no</Detailed_Variable></Parameters><ISO_Topic_Category>geoscientificInformation</ISO_Topic_Category><Keyword>Georg von Neumayer</Keyword><Keyword>GVN</Keyword><Keyword>Monitoring station</Keyword><Keyword>WCRP/GEWEX</Keyword><Sensor_Name><Long_Name>Ceilometer</Long_Name></Sensor_Name><Sensor_Name><Long_Name>Visibility sensor</Long_Name></Sensor_Name><Sensor_Name><Long_Name>Anemometer</Long_Name></Sensor_Name><Sensor_Name><Long_Name>Thermometer</Long_Name></Sensor_Name><Sensor_Name><Long_Name>Hygrometer</Long_Name></Sensor_Name><Sensor_Name><Long_Name>Barometer</Long_Name></Sensor_Name><Sensor_Name><Long_Name>Visual observation</Long_Name></Sensor_Name><Temporal_Coverage><Start_Date>1993-12-01</Start_Date><Stop_Date>1993-12-31</Stop_Date></Temporal_Coverage><Data_Set_Progress>Complete</Data_Set_Progress><Spatial_Coverage><Southernmost_Latitude>-70.65</Southernmost_Latitude><Northernmost_Latitude>-70.65</Northernmost_Latitude><Westernmost_Longitude>-8.25</Westernmost_Longitude><Easternmost_Longitude>-8.25</Easternmost_Longitude></Spatial_Coverage><Project><Short_Name>BSRN</Short_Name><Long_Name>Baseline Surface Radiation Network</Long_Name></Project><Access_Constraints>access rights needed</Access_Constraints><Data_Set_Language>English</Data_Set_Language><Originating_Center>Alfred Wegener Institute, Helmholtz Center for Polar and Marine Research, Bremerhaven</Originating_Center><Data_Center><Data_Center_Name><Short_Name>PANGAEA</Short_Name><Long_Name>Data Publisher for Earth &amp; Environmental Science</Long_Name></Data_Center_Name><Data_Center_URL>http://www.pangaea.de/</Data_Center_URL><Personnel><Role>Data Center Contact</Role><First_Name>Michael</First_Name><Last_Name>Diepenbroek</Last_Name><Email>mdiepenbroek@pangaea.de</Email><Phone>+49 421 218 65590</Phone><Contact_Address><Address>Leobener Str.</Address><City>Bremen</City><Province_or_State>Bremen</Province_or_State><Postal_Code>28359</Postal_Code><Country>Germany</Country></Contact_Address></Personnel></Data_Center><Distribution><Distribution_Media>online</Distribution_Media><Distribution_Size>3561 data points</Distribution_Size><Distribution_Format>text/tab-separated-values</Distribution_Format></Distribution><Reference>König-Langlo, Gert (2012): Validation routines for synoptic observations. Alfred Wegener Institute for Polar and Marine Research, Bremerhaven, Germany, hdl:10013/epic.40293.d001</Reference><Summary>For all details see the full metadata description at \"https://doi.pangaea.de/10.1594/PANGAEA.269656\"!</Summary><Related_URL><URL>http://bsrn.awi.de</URL><Description>BSRN</Description></Related_URL><Related_URL><URL>http://doi.pangaea.de/10013/epic.37965.d001</URL><Description>Characteristic of barometric tendency</Description></Related_URL><Related_URL><URL>http://doi.pangaea.de/10013/epic.37967.d001</URL><Description>High cloud</Description></Related_URL><Related_URL><URL>http://doi.pangaea.de/10013/epic.37968.d001</URL><Description>Low cloud</Description></Related_URL><Related_URL><URL>http://doi.pangaea.de/10013/epic.37969.d001</URL><Description>Middle cloud</Description></Related_URL><Related_URL><URL>http://doi.pangaea.de/10013/epic.37970.d001</URL><Description>Total cloud amount</Description></Related_URL><Related_URL><URL>http://doi.pangaea.de/10013/epic.37971.d001</URL><Description>Past weather1</Description></Related_URL><Related_URL><URL>http://doi.pangaea.de/10013/epic.37971.d001</URL><Description>Past weather2</Description></Related_URL><Related_URL><URL>http://doi.pangaea.de/10013/epic.37972.d001</URL><Description>Present weather</Description></Related_URL><Related_URL><URL>http://doi.pangaea.de/10013/epic.40293.d001</URL><Description>Validation routines for synoptic observations</Description></Related_URL><Related_URL><URL>http://en.wikipedia.org/wiki/ISO_8601</URL><Description>DATE/TIME</Description></Related_URL><Related_URL><URL>http://www.awi.de/en/about-us/organisation/staff/gert-koenig-langlo.html</URL></Related_URL><Related_URL><URL>http://www.awi.de/en/science/long-term-observations/atmosphere/antarctic-neumayer.html</URL><Description>GVN</Description></Related_URL><Related_URL><URL>http://www.weather.org.uk/resource/vvcode.htm</URL><Description>Horizontal visibility</Description></Related_URL><Metadata_Name>DIF</Metadata_Name><Metadata_Version>9.4</Metadata_Version><DIF_Creation_Date>2005-05-10</DIF_Creation_Date><Last_DIF_Revision_Date>2016-12-16</Last_DIF_Revision_Date></DIF></metadata></record></GetRecord></OAI-PMH>"
#> 

# curl options
library('httr')
pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.788382",
  config=verbose())

#> $`oai:pangaea.de:doi:10.1594/PANGAEA.788382`
#> $`oai:pangaea.de:doi:10.1594/PANGAEA.788382`$header
#> # A tibble: 1 × 3
#>                                  identifier            datestamp
#>                                       <chr>                <chr>
#> 1 oai:pangaea.de:doi:10.1594/PANGAEA.788382 2016-06-25T14:58:45Z
#> # ... with 1 more variables: setSpec <chr>
#> 
#> $`oai:pangaea.de:doi:10.1594/PANGAEA.788382`$metadata
#> # A tibble: 1 × 13
#>                                                                              title
#>                                                                              <chr>
#> 1 Trace metals in shells of mussels and clams from deep-sea hydrothermal vent fiel
#> # ... with 12 more variables: creator <chr>, source <chr>, publisher <chr>,
#> #   date <chr>, type <chr>, format <chr>, identifier <chr>, description <chr>,
#> #   language <chr>, rights <chr>, coverage <chr>, subject <chr>
#> 
#> 

# invalid record id
# pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.11111")
# pg_get_record(identifier = "oai:pangaea.de:doi:10.1594/PANGAEA.11111",
#   prefix="adfadf")