List identifiers of the Pangaea repository

List identifiers of the Pangaea repository

pg_list_identifiers(prefix = "oai_dc", from = NULL, until = NULL,
  set = NULL, token = NULL, as = "df", ...)

Arguments

prefix

A character string to specify the metadata format in OAI-PMH requests issued to the repository. The default (oai_dc) corresponds to the mandatory OAI unqualified Dublin Core metadata schema.

from

Character string giving datestamp to be used as lower bound for datestamp-based selective harvesting (i.e., only harvest records with datestamps in the given range). Dates and times must be encoded using ISO 8601. The trailing Z must be used when including time. OAI-PMH implies UTC for data/time specifications.

until

Character string giving a datestamp to be used as an upper bound, for datestamp-based selective harvesting (i.e., only harvest records with datestamps in the given range).

set

A character string giving a set to be used for selective harvesting (i.e., only harvest records in the given set).

token

(character) a token previously provided by the server to resume a request where it last left off. 50 is max number of records returned. We will loop for you internally to get all the records you asked for.

as

(character) What to return. One of "df" (for data.frame; default), "list", or "raw" (raw text)

...

Curl debugging options passed on to httr::GET()

Value

XML character string, data.frame, or list, depending on what requested with the as parameter

References

https://www.openarchives.org/pmh/

See also

wraps oai::list_identifiers()

Other oai methods: pg_get_record, pg_identify, pg_list_metadata_formats, pg_list_records, pg_list_sets

Examples

pg_list_identifiers(
  from = paste0(Sys.Date() - 2, "T00:00:00Z"),
  until = paste0(Sys.Date() - 1, "T18:00:00Z")
)

#> # A tibble: 1,993 × 7
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1   oai:pangaea.de:doi:10.1594/PANGAEA.55417 2017-03-08T13:19:32Z citable
#> 2   oai:pangaea.de:doi:10.1594/PANGAEA.57910 2017-03-08T13:17:50Z  ORFOIS
#> 3   oai:pangaea.de:doi:10.1594/PANGAEA.77250 2017-03-08T13:19:59Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.149999 2017-03-07T09:22:22Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.208129 2017-03-08T13:19:16Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.314690 2017-03-07T11:47:26Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.351146 2017-03-08T13:18:19Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.611095 2017-03-07T11:01:44Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.693923 2017-03-08T13:17:24Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.716835 2017-03-07T10:52:54Z citable
#> # ... with 1,983 more rows, and 4 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, setSpec.3 <chr>, setSpec.4 <chr>

pg_list_identifiers(set="geocode1", from=Sys.Date()-1, until=Sys.Date())

#> # A tibble: 604 × 7
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1   oai:pangaea.de:doi:10.1594/PANGAEA.55417 2017-03-08T13:19:32Z citable
#> 2   oai:pangaea.de:doi:10.1594/PANGAEA.77250 2017-03-08T13:19:59Z citable
#> 3  oai:pangaea.de:doi:10.1594/PANGAEA.351146 2017-03-08T13:18:19Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.693923 2017-03-08T13:17:24Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.726410 2017-03-08T13:17:54Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.794530 2017-03-08T13:18:46Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.803266 2017-03-08T13:17:31Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.804819 2017-03-08T13:19:03Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.805538 2017-03-08T13:19:18Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.837756 2017-03-08T13:18:04Z citable
#> # ... with 594 more rows, and 4 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, setSpec.3 <chr>, setSpec.4 <chr>

pg_list_identifiers(prefix="iso19139", from=Sys.Date()-1, until=Sys.Date())

#> # A tibble: 6,378 × 9
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1   oai:pangaea.de:doi:10.1594/PANGAEA.55417 2017-03-08T13:19:32Z citable
#> 2   oai:pangaea.de:doi:10.1594/PANGAEA.57910 2017-03-08T13:17:50Z  ORFOIS
#> 3   oai:pangaea.de:doi:10.1594/PANGAEA.77250 2017-03-08T13:19:59Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.208129 2017-03-08T13:19:16Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.351146 2017-03-08T13:18:19Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.693923 2017-03-08T13:17:24Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.726410 2017-03-08T13:17:54Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.757616 2017-03-08T13:18:05Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.758696 2017-03-08T13:19:02Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.760166 2017-03-09T14:54:46Z    BCCR
#> # ... with 6,368 more rows, and 6 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, setSpec.3 <chr>, setSpec.4 <chr>, setSpec.5 <chr>,
#> #   setSpec.6 <chr>

pg_list_identifiers(prefix="dif",
  from = paste0(Sys.Date() - 2, "T00:00:00Z"),
  until = paste0(Sys.Date() - 1, "T18:00:00Z")
)

#> # A tibble: 1,993 × 7
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1   oai:pangaea.de:doi:10.1594/PANGAEA.55417 2017-03-08T13:19:32Z citable
#> 2   oai:pangaea.de:doi:10.1594/PANGAEA.57910 2017-03-08T13:17:50Z  ORFOIS
#> 3   oai:pangaea.de:doi:10.1594/PANGAEA.77250 2017-03-08T13:19:59Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.149999 2017-03-07T09:22:22Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.208129 2017-03-08T13:19:16Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.314690 2017-03-07T11:47:26Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.351146 2017-03-08T13:18:19Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.611095 2017-03-07T11:01:44Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.693923 2017-03-08T13:17:24Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.716835 2017-03-07T10:52:54Z citable
#> # ... with 1,983 more rows, and 4 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, setSpec.3 <chr>, setSpec.4 <chr>