List records from Pangaea

List records from Pangaea

pg_list_records(prefix = "oai_dc", from = NULL, until = NULL,
  set = NULL, token = NULL, as = "df", ...)

Arguments

prefix

A character string to specify the metadata format in OAI-PMH requests issued to the repository. The default (oai_dc) corresponds to the mandatory OAI unqualified Dublin Core metadata schema.

from

Character string giving datestamp to be used as lower bound for datestamp-based selective harvesting (i.e., only harvest records with datestamps in the given range). Dates and times must be encoded using ISO 8601. The trailing Z must be used when including time. OAI-PMH implies UTC for data/time specifications.

until

Character string giving a datestamp to be used as an upper bound, for datestamp-based selective harvesting (i.e., only harvest records with datestamps in the given range).

set

A character string giving a set to be used for selective harvesting (i.e., only harvest records in the given set).

token

(character) a token previously provided by the server to resume a request where it last left off. 50 is max number of records returned. We will loop for you internally to get all the records you asked for.

as

(character) What to return. One of "df" (for data.frame; default), "list", or "raw" (raw text)

...

Curl debugging options passed on to httr::GET()

Value

XML character string, data.frame, or list, depending on what requested witht the as parameter

References

https://www.openarchives.org/pmh/

See also

wraps oai::list_records()

Other oai methods: pg_get_record, pg_identify, pg_list_identifiers, pg_list_metadata_formats, pg_list_sets

Examples

pg_list_records(set='citable', from=Sys.Date()-1, until=Sys.Date())

#> # A tibble: 142 × 44
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1   oai:pangaea.de:doi:10.1594/PANGAEA.55417 2017-03-08T13:19:32Z citable
#> 2   oai:pangaea.de:doi:10.1594/PANGAEA.57910 2017-03-08T13:17:50Z  ORFOIS
#> 3   oai:pangaea.de:doi:10.1594/PANGAEA.77250 2017-03-08T13:19:59Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.208129 2017-03-08T13:19:16Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.351146 2017-03-08T13:18:19Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.693923 2017-03-08T13:17:24Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.726410 2017-03-08T13:17:54Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.757616 2017-03-08T13:18:05Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.758696 2017-03-08T13:19:02Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.760166 2017-03-09T14:54:46Z    BCCR
#> # ... with 132 more rows, and 41 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, title <chr>, creator <chr>, creator.1 <chr>, source <chr>,
#> #   publisher <chr>, date <chr>, type <chr>, format <chr>, identifier.2 <chr>,
#> #   identifier.1 <chr>, description <chr>, language <chr>, rights <chr>,
#> #   rights.1 <chr>, coverage <chr>, subject <chr>, setSpec.3 <chr>,
#> #   creator.2 <chr>, creator.3 <chr>, relation <chr>, relation.1 <chr>,
#> #   relation.2 <chr>, relation.3 <chr>, relation.4 <chr>, creator.4 <chr>,
#> #   creator.5 <chr>, creator.6 <chr>, creator.7 <chr>, setSpec.4 <chr>,
#> #   setSpec.5 <chr>, setSpec.6 <chr>, source.1 <chr>, creator.8 <chr>,
#> #   creator.9 <chr>, creator.10 <chr>, creator.11 <chr>, creator.12 <chr>,
#> #   creator.13 <chr>, relation.5 <chr>

# When no results found > "'noRecordsMatch'"
# pg_list_records(set='geomound', from='2015-01-01', until='2015-01-01')

pg_list_records(prefix="iso19139", set='citable', from=Sys.Date()-1,
  until=Sys.Date())

#> # A tibble: 142 × 136
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1   oai:pangaea.de:doi:10.1594/PANGAEA.55417 2017-03-08T13:19:32Z citable
#> 2   oai:pangaea.de:doi:10.1594/PANGAEA.57910 2017-03-08T13:17:50Z  ORFOIS
#> 3   oai:pangaea.de:doi:10.1594/PANGAEA.77250 2017-03-08T13:19:59Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.208129 2017-03-08T13:19:16Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.351146 2017-03-08T13:18:19Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.693923 2017-03-08T13:17:24Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.726410 2017-03-08T13:17:54Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.757616 2017-03-08T13:18:05Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.758696 2017-03-08T13:19:02Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.760166 2017-03-09T14:54:46Z    BCCR
#> # ... with 132 more rows, and 133 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, fileIdentifier <chr>, contact <chr>, dateStamp <chr>,
#> #   dataSetURI <chr>, identificationInfo <chr>, contentInfo <chr>,
#> #   contentInfo.1 <chr>, contentInfo.2 <chr>, contentInfo.3 <chr>,
#> #   contentInfo.4 <chr>, contentInfo.5 <chr>, contentInfo.6 <chr>,
#> #   contentInfo.7 <chr>, contentInfo.8 <chr>, contentInfo.9 <chr>,
#> #   contentInfo.10 <chr>, contentInfo.11 <chr>, contentInfo.12 <chr>,
#> #   contentInfo.13 <chr>, contentInfo.14 <chr>, contentInfo.15 <chr>,
#> #   contentInfo.16 <chr>, contentInfo.17 <chr>, contentInfo.18 <chr>,
#> #   contentInfo.19 <chr>, contentInfo.20 <chr>, contentInfo.21 <chr>,
#> #   contentInfo.22 <chr>, distributionInfo <chr>, dataQualityInfo <chr>,
#> #   setSpec.3 <chr>, contentInfo.23 <chr>, contentInfo.24 <chr>,
#> #   contentInfo.25 <chr>, contentInfo.26 <chr>, contentInfo.27 <chr>,
#> #   contentInfo.28 <chr>, contentInfo.29 <chr>, contentInfo.30 <chr>,
#> #   setSpec.4 <chr>, setSpec.5 <chr>, setSpec.6 <chr>, contentInfo.31 <chr>,
#> #   contentInfo.32 <chr>, contentInfo.33 <chr>, contentInfo.34 <chr>,
#> #   contentInfo.35 <chr>, contentInfo.36 <chr>, contentInfo.37 <chr>,
#> #   contentInfo.38 <chr>, contentInfo.39 <chr>, contentInfo.40 <chr>,
#> #   contentInfo.41 <chr>, contentInfo.42 <chr>, contentInfo.43 <chr>,
#> #   contentInfo.44 <chr>, contentInfo.45 <chr>, contentInfo.46 <chr>,
#> #   contentInfo.47 <chr>, contentInfo.48 <chr>, contentInfo.49 <chr>,
#> #   contentInfo.50 <chr>, contentInfo.51 <chr>, contentInfo.52 <chr>,
#> #   contentInfo.53 <chr>, contentInfo.54 <chr>, contentInfo.55 <chr>,
#> #   contentInfo.56 <chr>, contentInfo.57 <chr>, contentInfo.58 <chr>,
#> #   contentInfo.59 <chr>, contentInfo.60 <chr>, contentInfo.61 <chr>,
#> #   contentInfo.62 <chr>, contentInfo.63 <chr>, contentInfo.64 <chr>,
#> #   contentInfo.65 <chr>, contentInfo.66 <chr>, contentInfo.67 <chr>,
#> #   contentInfo.68 <chr>, contentInfo.69 <chr>, contentInfo.70 <chr>,
#> #   contentInfo.71 <chr>, contentInfo.72 <chr>, contentInfo.73 <chr>,
#> #   contentInfo.74 <chr>, contentInfo.75 <chr>, contentInfo.76 <chr>,
#> #   contentInfo.77 <chr>, contentInfo.78 <chr>, contentInfo.79 <chr>,
#> #   contentInfo.80 <chr>, contentInfo.81 <chr>, contentInfo.82 <chr>,
#> #   contentInfo.83 <chr>, contentInfo.84 <chr>, contentInfo.85 <chr>,
#> #   contentInfo.86 <chr>, ...

## FIXME - below are broken
# pg_list_records(prefix="dif", set='citable', from=Sys.Date()-4,
#   until=Sys.Date())
# pg_list_records(prefix="dif", set='project4094', from=Sys.Date()-4,
#   until=Sys.Date())