Download data from Pangaea.

Grabs data as a dataframe or list of dataframes from a Pangaea data repository URI; see: https://www.pangaea.de/

pg_cache_clear(doi = NULL, prompt = TRUE)

pg_cache_list()

pg_data(doi, overwrite = TRUE, verbose = TRUE, ...)

Arguments

doi

DOI of Pangaeae single dataset, or of a collection of datasets. Expects either just a DOI of the form 10.1594/PANGAEA.746398, or with the URL part in front, like https://doi.pangaea.de/10.1594/PANGAEA.746398

prompt

(logical) Prompt before clearing all files in cache? No prompt used when DOIs assed in. Default: TRUE

overwrite

(logical) Ovewrite a file if one is found with the same name

verbose

(logical) print information messages. Default: TRUE

...

Curl options passed on to httr::GET()

Value

One or more items of class pangaea, each with a citation object, metadata object, and data object. Each data object is printed as a tbl_df object, but the actual object is simply a data.frame.

Details

Data files are stored in an operating system appropriate location. Run rappdirs::user_cache_dir("pangaear") to get the storage location on your machine.

References

https://www.pangaea.de

Examples

# a single file
res <- pg_data(doi='10.1594/PANGAEA.807580')

#> Downloading 1 datasets from 10.1594/PANGAEA.807580

#> Processing 1 files

res

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.807580
#> # A tibble: 32,179 × 13
#>                Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>                <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 2  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 3  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 4  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 5  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 6  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 7  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 8  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 9  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 10 M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> # ... with 32,169 more rows, and 8 more variables: `Depth water [m]` <dbl>,
#> #   `Press [dbar]` <int>, `Temp [°C]` <dbl>, Sal <dbl>, `Tpot [°C]` <dbl>,
#> #   `Sigma-theta [kg/m**3]` <dbl>, `Sigma in situ [kg/m**3]` <dbl>, `Cond
#> #   [mS/cm]` <dbl>
#> 

res[[1]]$doi

#> [1] "10.1594/PANGAEA.807580"

res[[1]]$citation

#> See https://doi.pangaea.de/10.1594/PANGAEA.807580 for the citation

res[[1]]$meta

#> Citation:	Schiebel, Ralf; Waniek, Joanna J; Bork, Matthias; Hemleben, Christoph (2001): Physical oceanography during METEOR cruise M36/6. doi:10.1594/PANGAEA.807580, 
#> 	In supplement to: Schiebel, R et al. (2001): Planktic foraminiferal production stimulated by chlorophyll redistribution and entrainment of nutrients. Deep Sea Research Part I: Oceanographic Research Papers, 48(3), 721-740, doi:10.1016/S0967-0637(00)00065-0
#> Project(s):	Joint Global Ocean Flux Study (JGOFS) (URI: http://ijgofs.whoi.edu/)
#> Coverage:	MEDIAN LATITUDE: 47.554050 * MEDIAN LONGITUDE: -18.229529 * SOUTH-BOUND LATITUDE: 45.598300 * WEST-BOUND LONGITUDE: -19.581700 * NORTH-BOUND LATITUDE: 48.968300 * EAST-BOUND LONGITUDE: -16.461700
#> 	DATE/TIME START: 1996-10-14T12:24:00 * DATE/TIME END: 1996-11-01T07:18:00
#> 	MINIMUM DEPTH, water: 0.00 m * MAXIMUM DEPTH, water: 4822.01 m
#> Event(s):	M36/6-CTD-365_003 * LATITUDE: 48.966700 * LONGITUDE: -16.463300 * DATE/TIME START: 1996-10-14T12:24:00 * DATE/TIME END: 1996-10-14T13:11:00 * ELEVATION: -4802.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-366_004 * LATITUDE: 48.966700 * LONGITUDE: -16.478300 * DATE/TIME START: 1996-10-16T09:40:00 * DATE/TIME END: 1996-10-16T11:39:00 * ELEVATION: -4802.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-375_005 * LATITUDE: 48.966700 * LONGITUDE: -16.461700 * DATE/TIME START: 1996-10-19T16:55:00 * DATE/TIME END: 1996-10-19T18:48:00 * ELEVATION: -4802.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-380_006 * LATITUDE: 48.968300 * LONGITUDE: -16.466700 * DATE/TIME START: 1996-10-20T22:52:00 * DATE/TIME END: 1996-10-21T00:46:00 * ELEVATION: -4804.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-386_007 * LATITUDE: 47.181700 * LONGITUDE: -19.561700 * DATE/TIME START: 1996-10-23T12:41:00 * DATE/TIME END: 1996-10-23T13:26:00 * ELEVATION: -4565.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-392_008 * LATITUDE: 47.196700 * LONGITUDE: -19.581700 * DATE/TIME START: 1996-10-24T20:21:00 * DATE/TIME END: 1996-10-24T22:16:00 * ELEVATION: -4561.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-393_009 * LATITUDE: 47.588300 * LONGITUDE: -18.861700 * DATE/TIME START: 1996-10-25T02:40:00 * DATE/TIME END: 1996-10-25T03:16:00 * ELEVATION: -4120.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-394_010 * LATITUDE: 48.095000 * LONGITUDE: -18.156700 * DATE/TIME START: 1996-10-25T08:27:00 * DATE/TIME END: 1996-10-25T09:05:00 * ELEVATION: -4436.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-397_011 * LATITUDE: 47.178300 * LONGITUDE: -19.568300 * DATE/TIME START: 1996-10-28T12:39:00 * DATE/TIME END: 1996-10-28T13:16:00 * ELEVATION: -4567.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-404_012 * LATITUDE: 47.186700 * LONGITUDE: -19.560000 * DATE/TIME START: 1996-10-30T08:01:00 * DATE/TIME END: 1996-10-30T09:49:00 * ELEVATION: -4565.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-407_013 * LATITUDE: 47.183300 * LONGITUDE: -19.566700 * DATE/TIME START: 1996-10-31T03:42:00 * DATE/TIME END: 1996-10-31T04:55:00 * ELEVATION: -4566.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-409_014 * LATITUDE: 46.585000 * LONGITUDE: -18.873300 * DATE/TIME START: 1996-10-31T20:32:00 * DATE/TIME END: 1996-10-31T21:07:00 * ELEVATION: -4610.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-410_015 * LATITUDE: 46.095000 * LONGITUDE: -18.155000 * DATE/TIME START: 1996-11-01T01:25:00 * DATE/TIME END: 1996-11-01T02:00:00 * ELEVATION: -4120.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> 	M36/6-CTD-411_016 * LATITUDE: 45.598300 * LONGITUDE: -17.458300 * DATE/TIME START: 1996-11-01T06:45:00 * DATE/TIME END: 1996-11-01T07:18:00 * ELEVATION: -4651.0 m * CAMPAIGN: M36/6 (URI: doi:10.2312/cr_m36) * BASIS: Meteor (1986) (URI: https://de.wikipedia.org/wiki/Meteor_(Schiff,_1986)) * DEVICE: CTD/Rosette (CTD-RO)
#> Parameter(s):	Event label (Event)
#> 	Date/Time of event (Date/Time)
#> 	Latitude of event (Latitude)
#> 	Longitude of event (Longitude)
#> 	Elevation of event [m] (Elevation)
#> 	DEPTH, water [m] (Depth water) * GEOCODE
#> 	Pressure, water [dbar] (Press) * PI: Waniek, Joanna J (joanna.waniek@io-warnemuende.de, http://www.io-warnemuende.de/homepages/waniek/) * METHOD: CTD, Neil Brown, Mark III B
#> 	Temperature, water [°C] (Temp) * PI: Waniek, Joanna J (joanna.waniek@io-warnemuende.de, http://www.io-warnemuende.de/homepages/waniek/) * METHOD: CTD, Neil Brown, Mark III B
#> 	Salinity (Sal) * PI: Waniek, Joanna J (joanna.waniek@io-warnemuende.de, http://www.io-warnemuende.de/homepages/waniek/) * METHOD: Calculated from conductivity
#> 	Temperature, water, potential [°C] (Tpot) * PI: Waniek, Joanna J (joanna.waniek@io-warnemuende.de, http://www.io-warnemuende.de/homepages/waniek/) * METHOD: Calculated
#> 	Density, sigma-theta (0) [kg/m**3] (Sigma-theta) * PI: Waniek, Joanna J (joanna.waniek@io-warnemuende.de, http://www.io-warnemuende.de/homepages/waniek/) * METHOD: Calculated
#> 	Density, sigma, in situ [kg/m**3] (Sigma in situ) * PI: Waniek, Joanna J (joanna.waniek@io-warnemuende.de, http://www.io-warnemuende.de/homepages/waniek/) * METHOD: Calculated
#> 	Conductivity [mS/cm] (Cond) * PI: Waniek, Joanna J (joanna.waniek@io-warnemuende.de, http://www.io-warnemuende.de/homepages/waniek/) * METHOD: CTD, Neil Brown, Mark III B
#> License:	Creative Commons Attribution 3.0 Unported (CC-BY)
#> Size:	225253 data points

res[[1]]$data

#> # A tibble: 32,179 × 13
#>                Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>                <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 2  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 3  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 4  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 5  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 6  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 7  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 8  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 9  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 10 M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> # ... with 32,169 more rows, and 8 more variables: `Depth water [m]` <dbl>,
#> #   `Press [dbar]` <int>, `Temp [°C]` <dbl>, Sal <dbl>, `Tpot [°C]` <dbl>,
#> #   `Sigma-theta [kg/m**3]` <dbl>, `Sigma in situ [kg/m**3]` <dbl>, `Cond
#> #   [mS/cm]` <dbl>

# another single file
pg_data(doi='10.1594/PANGAEA.807584')

#> Downloading 1 datasets from 10.1594/PANGAEA.807584

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.807584
#> # A tibble: 698 × 53
#>            Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>            <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 2  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 3  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 4  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 5  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 6  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 7  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 8  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 9  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 10 M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> # ... with 688 more rows, and 48 more variables: `Depth water [m]` <int>,
#> #   `Depth top [m]` <int>, `Depth bot [m]` <int>, Size <chr>, `D. anfracta
#> #   [#]` <int>, `G. bulloides [#]` <int>, `G. calida [#]` <int>, `T. clarkei
#> #   [#]` <int>, `T. compressa [#]` <int>, `G. conglobatus [#]` <int>, `G.
#> #   crassaformis [#]` <int>, `S. dehiscens [#]` <int>, `B. digitata [#]` <int>,
#> #   `G. digitata [#]` <int>, `H. digitata [#]` <int>, `N. dutertrei [#]` <int>,
#> #   `G. falconensis [#]` <int>, `S. globigerum [#]` <int>, `G. glutinata
#> #   [#]` <int>, `G. hirsuta [#]` <int>, `T. humilis [#]` <int>, `N. incompta
#> #   [#]` <int>, `G. inflata [#]` <int>, `T. iota [#]` <int>, `G. menardii
#> #   [#]` <int>, `G. minuta [#]` <int>, `C. nitida [#]` <int>, `P.
#> #   obliquiloculata [#]` <int>, `N. pachyderma [#]` <int>, `T. parkerae
#> #   [#]` <int>, `H. pelagica [#]` <int>, `B. pumilio [#]` <int>, `T.
#> #   quinqueloba [#]` <int>, `G. ruber p [#]` <int>, `G. ruber w [#]` <int>, `G.
#> #   rubescens [#]` <int>, `G. sacculifer [#]` <int>, `G. scitula [#]` <int>,
#> #   `G. siphonifera [#]` <int>, `G. tenella [#]` <int>, `G. theyeri [#]` <int>,
#> #   `G. truncatulinoides [#]` <int>, `G. tumida [#]` <int>, `G. ungulata
#> #   [#]` <int>, `O. universa [#]` <int>, `G. uvula [#]` <int>, `G. vivans
#> #   [#]` <int>, `Foram plankt indet [#]` <int>
#> 

# Many files
(res <- pg_data(doi='10.1594/PANGAEA.761032'))

#> Downloading 2 datasets from 10.1594/PANGAEA.761032

#> Processing 2 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.761002
#> # A tibble: 250,201 × 9
#>          Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>          <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 2  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 3  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 4  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 5  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 6  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 7  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 8  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 9  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 10 NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> # ... with 250,191 more rows, and 4 more variables: `Depth water [m]` <int>,
#> #   `Temp [°C]` <dbl>, Sal <dbl>, `Tpot [°C]` <dbl>
#> 
#> [[2]]
#> <Pangaea data> 10.1594/PANGAEA.319391
#> # A tibble: 6,060 × 12
#>         `Date/Time` Latitude Longitude `Depth water [m]` `UC [cm/s]`
#>               <chr>    <dbl>     <dbl>             <int>       <dbl>
#> 1  1997-08-03T19:00 -56.3617  -50.6079                40        53.8
#> 2  1997-08-03T19:00 -56.3617  -50.6079                50        51.3
#> 3  1997-08-03T19:00 -56.3617  -50.6079                60        50.9
#> 4  1997-08-03T19:00 -56.3617  -50.6079                70        51.0
#> 5  1997-08-03T19:00 -56.3617  -50.6079                80        49.8
#> 6  1997-08-03T19:00 -56.3617  -50.6079                90        49.5
#> 7  1997-08-03T19:00 -56.3617  -50.6079               100        47.7
#> 8  1997-08-03T19:00 -56.3617  -50.6079               110        48.3
#> 9  1997-08-03T19:00 -56.3617  -50.6079               120        48.9
#> 10 1997-08-03T19:00 -56.3617  -50.6079               130        48.7
#> # ... with 6,050 more rows, and 7 more variables: `VC [cm/s]` <dbl>, `T tech
#> #   [°C]` <dbl>, `T tech std dev [±]` <dbl>, `Ship velo SU [m/s]` <dbl>, `Ship
#> #   velo SU std dev [±]` <dbl>, `Ship velo SV [m/s]` <dbl>, `Ship velo SV std
#> #   dev [±]` <dbl>
#> 

res[[1]]

#> <Pangaea data> 10.1594/PANGAEA.761002
#> # A tibble: 250,201 × 9
#>          Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>          <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 2  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 3  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 4  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 5  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 6  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 7  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 8  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 9  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 10 NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> # ... with 250,191 more rows, and 4 more variables: `Depth water [m]` <int>,
#> #   `Temp [°C]` <dbl>, Sal <dbl>, `Tpot [°C]` <dbl>

res[[2]]

#> <Pangaea data> 10.1594/PANGAEA.319391
#> # A tibble: 6,060 × 12
#>         `Date/Time` Latitude Longitude `Depth water [m]` `UC [cm/s]`
#>               <chr>    <dbl>     <dbl>             <int>       <dbl>
#> 1  1997-08-03T19:00 -56.3617  -50.6079                40        53.8
#> 2  1997-08-03T19:00 -56.3617  -50.6079                50        51.3
#> 3  1997-08-03T19:00 -56.3617  -50.6079                60        50.9
#> 4  1997-08-03T19:00 -56.3617  -50.6079                70        51.0
#> 5  1997-08-03T19:00 -56.3617  -50.6079                80        49.8
#> 6  1997-08-03T19:00 -56.3617  -50.6079                90        49.5
#> 7  1997-08-03T19:00 -56.3617  -50.6079               100        47.7
#> 8  1997-08-03T19:00 -56.3617  -50.6079               110        48.3
#> 9  1997-08-03T19:00 -56.3617  -50.6079               120        48.9
#> 10 1997-08-03T19:00 -56.3617  -50.6079               130        48.7
#> # ... with 6,050 more rows, and 7 more variables: `VC [cm/s]` <dbl>, `T tech
#> #   [°C]` <dbl>, `T tech std dev [±]` <dbl>, `Ship velo SU [m/s]` <dbl>, `Ship
#> #   velo SU std dev [±]` <dbl>, `Ship velo SV [m/s]` <dbl>, `Ship velo SV std
#> #   dev [±]` <dbl>

# Manipulating the cache
## list files in the cache
pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_804606.txt"
#> [13] "10_1594_PANGAEA_807580.txt" "10_1594_PANGAEA_807584.txt"

## clear all data
# pg_cache_clear()
pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_804606.txt"
#> [13] "10_1594_PANGAEA_807580.txt" "10_1594_PANGAEA_807584.txt"

## clear a single dataset by DOI
pg_data(doi='10.1594/PANGAEA.812093')

#> Downloading 1 datasets from 10.1594/PANGAEA.812093

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.812093
#> # A tibble: 36 × 18
#>       Event `Sample label`                              `Sample comment`
#>       <chr>          <chr>                                         <chr>
#> 1  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 2  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 3  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 4  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 5  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 6  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 7  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 8  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 9  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 10 169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> # ... with 26 more rows, and 15 more variables: `Depth [m]` <dbl>, `Distance
#> #   [µm]` <int>, `T tech [°C]` <dbl>, Run <chr>, `- (initial concentration of
#> #   diff...)` <dbl>, `- (initial concentration of diff...)` <dbl>, `Poros frac
#> #   (#1, from fit)` <dbl>, `Poros frac (#2, from fit)` <dbl>, `Diff coeff
#> #   [10**-6 cm**2/s]` <dbl>, `Diff coeff std dev [±] (effective from
#> #   fit)` <dbl>, `Diff coeff std dev [±] (effective total)` <dbl>, X <dbl>, `X
#> #   (average)` <int>, F <int>, `- (m)` <dbl>
#> 

pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_804606.txt"
#> [13] "10_1594_PANGAEA_807580.txt" "10_1594_PANGAEA_807584.txt"
#> [15] "10_1594_PANGAEA_812093.txt"

pg_cache_clear(doi='10.1594/PANGAEA.812093')
pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_804606.txt"
#> [13] "10_1594_PANGAEA_807580.txt" "10_1594_PANGAEA_807584.txt"

## clear more than 1 dataset by DOI
lapply(c('10.1594/PANGAEA.746398','10.1594/PANGAEA.746400'), pg_data)

#> Downloading 1 datasets from 10.1594/PANGAEA.746398

#> Processing 1 files

#> Downloading 1 datasets from 10.1594/PANGAEA.746400

#> Processing 1 files

#> [[1]]
#> [[1]][[1]]
#> <Pangaea data> 10.1594/PANGAEA.746398
#> # A tibble: 4 × 13
#>   `Depth [m]` `Depth top [m]` `Depth bot [m]`         Sediment `1-0.5 mm [%]`
#>         <dbl>           <dbl>           <dbl>            <chr>          <dbl>
#> 1       0.025            0.00            0.05 coccolithic ooze           0.44
#> 2       0.160            0.15            0.17 coccolithic ooze           5.05
#> 3       0.175            0.35            0.00 coccolithic ooze           2.85
#> 4       0.205            0.17            0.24 coccolithic ooze           8.12
#> # ... with 8 more variables: `500-250 µm [%]` <dbl>, `250-100 µm [%]` <dbl>,
#> #   `100-50 µm [%]` <dbl>, `50-10 µm [%]` <dbl>, `10-5 µm [%]` <dbl>, `5-1 µm
#> #   [%]` <dbl>, `<1 µm [%]` <dbl>, `Sum [%]` <dbl>
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> <Pangaea data> 10.1594/PANGAEA.746400
#> # A tibble: 11 × 13
#>    `Depth [m]` `Depth top [m]` `Depth bot [m]`           Sediment
#>          <dbl>           <dbl>           <dbl>              <chr>
#> 1        0.025            0.00            0.05 foraminiferal ooze
#> 2        0.025            0.00            0.05 foraminiferal ooze
#> 3        0.375            0.35            0.40 foraminiferal ooze
#> 4        0.375            0.35            0.40 foraminiferal ooze
#> 5        0.475            0.45            0.50   coccolithic ooze
#> 6        1.025            1.00            1.05 foraminiferal ooze
#> 7        1.025            1.00            1.05 foraminiferal ooze
#> 8        1.725            1.70            1.75 foraminiferal ooze
#> 9        1.725            1.70            1.75 foraminiferal ooze
#> 10       2.475            2.45            2.50   coccolithic ooze
#> 11       3.105            3.08            3.13   coccolithic ooze
#> # ... with 9 more variables: `1-0.5 mm [%]` <dbl>, `500-250 µm [%]` <dbl>,
#> #   `250-100 µm [%]` <dbl>, `100-50 µm [%]` <dbl>, `50-10 µm [%]` <dbl>, `10-5
#> #   µm [%]` <dbl>, `5-1 µm [%]` <dbl>, `<1 µm [%]` <dbl>, `Sum [%]` <dbl>
#> 
#> 

pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_746398.txt"
#>  [3] "10_1594_PANGAEA_746400.txt" "10_1594_PANGAEA_761002.txt"
#>  [5] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [7] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [9] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#> [11] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [13] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_804606.txt"
#> [15] "10_1594_PANGAEA_807580.txt" "10_1594_PANGAEA_807584.txt"

pg_cache_clear(doi=c('10.1594/PANGAEA.746398','10.1594/PANGAEA.746400'))
pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_804606.txt"
#> [13] "10_1594_PANGAEA_807580.txt" "10_1594_PANGAEA_807584.txt"

# search for datasets, then pass in DOIs
res <- pg_search(query='water', count=20)
pg_data(res$doi[1])

#> Downloading 7 datasets from 10.1594/PANGAEA.843740

#> Processing 7 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.843733
#> # A tibble: 52,560 × 5
#>         `Date/Time` `Press [dbar]`
#>               <chr>          <dbl>
#> 1  2005-01-01T00:00          12.70
#> 2  2005-01-01T00:10          12.79
#> 3  2005-01-01T00:20          12.87
#> 4  2005-01-01T00:30          12.92
#> 5  2005-01-01T00:40          12.99
#> 6  2005-01-01T00:50          13.05
#> 7  2005-01-01T01:00          13.11
#> 8  2005-01-01T01:10          13.16
#> 9  2005-01-01T01:20          13.19
#> 10 2005-01-01T01:30          13.22
#> # ... with 52,550 more rows, and 3 more variables: `Water level [m] (with trend
#> #   and outlier removal)` <dbl>, `Water level [m] (after all processing
#> #   steps)` <dbl>, `Flag [#]` <int>
#> 
#> [[2]]
#> <Pangaea data> 10.1594/PANGAEA.843734
#> # A tibble: 52,560 × 5
#>         `Date/Time` `Press [dbar]`
#>               <chr>          <dbl>
#> 1  2006-01-01T00:00          12.09
#> 2  2006-01-01T00:10          12.00
#> 3  2006-01-01T00:20          11.90
#> 4  2006-01-01T00:30          11.79
#> 5  2006-01-01T00:40          11.69
#> 6  2006-01-01T00:50          11.59
#> 7  2006-01-01T01:00          11.48
#> 8  2006-01-01T01:10          11.37
#> 9  2006-01-01T01:20          11.26
#> 10 2006-01-01T01:30          11.15
#> # ... with 52,550 more rows, and 3 more variables: `Water level [m] (with trend
#> #   and outlier removal)` <dbl>, `Water level [m] (after all processing
#> #   steps)` <dbl>, `Flag [#]` <int>
#> 
#> [[3]]
#> <Pangaea data> 10.1594/PANGAEA.843735
#> # A tibble: 52,560 × 5
#>         `Date/Time` `Press [dbar]`
#>               <chr>          <dbl>
#> 1  2007-01-01T00:00          11.41
#> 2  2007-01-01T00:10          11.39
#> 3  2007-01-01T00:20          11.34
#> 4  2007-01-01T00:30          11.28
#> 5  2007-01-01T00:40          11.31
#> 6  2007-01-01T00:50          11.32
#> 7  2007-01-01T01:00          11.29
#> 8  2007-01-01T01:10          11.23
#> 9  2007-01-01T01:20          11.29
#> 10 2007-01-01T01:30          11.31
#> # ... with 52,550 more rows, and 3 more variables: `Water level [m] (with trend
#> #   and outlier removal)` <dbl>, `Water level [m] (after all processing
#> #   steps)` <dbl>, `Flag [#]` <int>
#> 
#> [[4]]
#> <Pangaea data> 10.1594/PANGAEA.843736
#> # A tibble: 52,704 × 5
#>         `Date/Time` `Press [dbar]`
#>               <chr>          <dbl>
#> 1  2008-01-01T00:00          11.40
#> 2  2008-01-01T00:10          11.55
#> 3  2008-01-01T00:20          11.67
#> 4  2008-01-01T00:30          11.80
#> 5  2008-01-01T00:40          11.92
#> 6  2008-01-01T00:50          12.01
#> 7  2008-01-01T01:00          12.10
#> 8  2008-01-01T01:10          12.18
#> 9  2008-01-01T01:20          12.27
#> 10 2008-01-01T01:30          12.34
#> # ... with 52,694 more rows, and 3 more variables: `Water level [m] (with trend
#> #   and outlier removal)` <dbl>, `Water level [m] (after all processing
#> #   steps)` <dbl>, `Flag [#]` <int>
#> 
#> [[5]]
#> <Pangaea data> 10.1594/PANGAEA.843737
#> # A tibble: 52,560 × 5
#>         `Date/Time` `Press [dbar]`
#>               <chr>          <dbl>
#> 1  2009-01-01T00:00          11.63
#> 2  2009-01-01T00:10          11.70
#> 3  2009-01-01T00:20          11.78
#> 4  2009-01-01T00:30          11.84
#> 5  2009-01-01T00:40          11.90
#> 6  2009-01-01T00:50          11.94
#> 7  2009-01-01T01:00          11.97
#> 8  2009-01-01T01:10          11.98
#> 9  2009-01-01T01:20          11.97
#> 10 2009-01-01T01:30          11.95
#> # ... with 52,550 more rows, and 3 more variables: `Water level [m] (with trend
#> #   and outlier removal)` <dbl>, `Water level [m] (after all processing
#> #   steps)` <dbl>, `Flag [#]` <int>
#> 
#> [[6]]
#> <Pangaea data> 10.1594/PANGAEA.843738
#> # A tibble: 52,560 × 5
#>         `Date/Time` `Press [dbar]`
#>               <chr>          <dbl>
#> 1  2010-01-01T00:00          12.12
#> 2  2010-01-01T00:10          12.02
#> 3  2010-01-01T00:20          11.90
#> 4  2010-01-01T00:30          11.79
#> 5  2010-01-01T00:40          11.69
#> 6  2010-01-01T00:50          11.58
#> 7  2010-01-01T01:00          11.48
#> 8  2010-01-01T01:10          11.37
#> 9  2010-01-01T01:20          11.26
#> 10 2010-01-01T01:30          11.16
#> # ... with 52,550 more rows, and 3 more variables: `Water level [m] (with trend
#> #   and outlier removal)` <dbl>, `Water level [m] (after all processing
#> #   steps)` <dbl>, `Flag [#]` <int>
#> 
#> [[7]]
#> <Pangaea data> 10.1594/PANGAEA.843739
#> # A tibble: 52,560 × 5
#>         `Date/Time` `Press [dbar]`
#>               <chr>          <dbl>
#> 1  2011-01-01T00:00          11.11
#> 2  2011-01-01T00:10          11.07
#> 3  2011-01-01T00:20          11.03
#> 4  2011-01-01T00:30          11.01
#> 5  2011-01-01T00:40          10.97
#> 6  2011-01-01T00:50          10.93
#> 7  2011-01-01T01:00          10.94
#> 8  2011-01-01T01:10          10.93
#> 9  2011-01-01T01:20          10.93
#> 10 2011-01-01T01:30          10.94
#> # ... with 52,550 more rows, and 3 more variables: `Water level [m] (with trend
#> #   and outlier removal)` <dbl>, `Water level [m] (after all processing
#> #   steps)` <dbl>, `Flag [#]` <int>
#> 

pg_data(res$doi[2])

#> Downloading 1 datasets from 10.1594/PANGAEA.762952

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.762952
#> # A tibble: 13 × 10
#>                    Species               `Growth form` `Water dm [%] (initial)`
#>                      <chr>                       <chr>                    <int>
#> 1  Schistidium antarcticum              entire cushion                      133
#> 2  Schistidium antarcticum                   tall turf                      554
#> 3  Schistidium antarcticum                  short turf                      565
#> 4             Bryum algens  shorl turf, dense rhizoids                      717
#> 5             Bryum algens short turf, sparse rhizoids                      182
#> 6         Usnca sphacelata                 robust form                       73
#> 7         Usnca sphacelata                   fine form                      114
#> 8         Usnca sphacelata                  shade form                      103
#> 9         Usnea antarctica                 robust form                       85
#> 10        Usnea antarctica                   fine form                       83
#> 11     Alectoria minuscula                                                  351
#> 12   Umbilicaria decussata                                                  130
#> 13      Umbilicaria aprina                                                  204
#> # ... with 7 more variables: `Water std dev [±] (initial)` <int>, `Water loss
#> #   dry wgt [%] (Time 1 hr)` <int>, `Water loss dry wgt [%] (Time 6 hr)` <int>,
#> #   `Water loss dry wgt [%] (Time 12 hr)` <int>, `Water loss dry wgt [%] (Time
#> #   24 hr)` <int>, `Water dm [%]` <int>, `Water std dev [±]` <dbl>
#> 

pg_data(res$doi[3])

#> Downloading 10 datasets from 10.1594/PANGAEA.811079

#> Processing 10 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.811075
#> # A tibble: 2,165 × 2
#>         `Date/Time` `Q [l/s]`
#>               <chr>     <dbl>
#> 1  2011-07-04T02:15    1.0264
#> 2  2011-07-04T02:45    0.9767
#> 3  2011-07-04T03:15    1.0193
#> 4  2011-07-04T03:45    0.9557
#> 5  2011-07-04T04:15    0.9879
#> 6  2011-07-04T04:45    1.0351
#> 7  2011-07-04T05:15    1.0956
#> 8  2011-07-04T05:45    1.1674
#> 9  2011-07-04T06:15    1.1720
#> 10 2011-07-04T06:45    1.1856
#> # ... with 2,155 more rows
#> 
#> [[2]]
#> <Pangaea data> 10.1594/PANGAEA.811076
#> # A tibble: 4,215 × 3
#>         `Date/Time` `ET [mm/h]` `Precip [mm/h]`
#>               <chr>       <dbl>           <dbl>
#> 1  2011-05-24T04:45      0.0116          0.0000
#> 2  2011-05-24T05:15      0.0149          0.0000
#> 3  2011-05-24T05:45      0.0182          0.4464
#> 4  2011-05-24T06:15      0.0230          0.0000
#> 5  2011-05-24T06:45      0.0278          0.0000
#> 6  2011-05-24T07:15      0.0174          0.0000
#> 7  2011-05-24T07:45      0.0140          0.0000
#> 8  2011-05-24T08:15      0.0096          0.0000
#> 9  2011-05-24T08:45      0.0135          0.0000
#> 10 2011-05-24T09:15      0.0158          0.0000
#> # ... with 4,205 more rows
#> 
#> [[3]]
#> <Pangaea data> 10.1594/PANGAEA.811084
#> # A tibble: 13 × 4
#>    `Date/Time` `Depth [m]` `Kfs [m/day]` `Soil comp`
#>          <chr>       <dbl>         <dbl>       <chr>
#> 1   2012-07-01        0.02        42.336        moss
#> 2   2012-07-02        0.07         3.283     organic
#> 3   2012-07-03        0.07         2.246     organic
#> 4   2012-07-04        0.14         0.207     mineral
#> 5   2012-07-05        0.23         0.030     mineral
#> 6   2012-07-06        0.23         0.199     mineral
#> 7   2012-07-07        0.02        38.016        moss
#> 8   2012-07-08        0.10         1.123     organic
#> 9   2012-07-09        0.18         0.458     mineral
#> 10  2012-07-10        0.17         0.214     mineral
#> 11  2012-07-11        0.17         0.049     mineral
#> 12  2012-07-12        0.10         0.562     organic
#> 13  2012-07-13        0.07        17.280     organic
#> 
#> [[4]]
#> <Pangaea data> 10.1594/PANGAEA.811070
#> # A tibble: 4,215 × 2
#>         `Date/Time` `NLF [mm/h]`
#>               <chr>        <dbl>
#> 1  2011-05-24T04:45      -0.2604
#> 2  2011-05-24T05:15      -0.2216
#> 3  2011-05-24T05:45       0.0010
#> 4  2011-05-24T06:15      -0.0697
#> 5  2011-05-24T06:45      -0.1706
#> 6  2011-05-24T07:15      -0.0451
#> 7  2011-05-24T07:45       0.0338
#> 8  2011-05-24T08:15      -0.0425
#> 9  2011-05-24T08:45      -0.0594
#> 10 2011-05-24T09:15      -0.2294
#> # ... with 4,205 more rows
#> 
#> [[5]]
#> <Pangaea data> 10.1594/PANGAEA.811071
#> # A tibble: 4,215 × 2
#>         `Date/Time` `NLF [mm/h]`
#>               <chr>        <dbl>
#> 1  2011-05-24T04:45       0.6314
#> 2  2011-05-24T05:15       0.3428
#> 3  2011-05-24T05:45      -0.2033
#> 4  2011-05-24T06:15       0.1897
#> 5  2011-05-24T06:45       0.0846
#> 6  2011-05-24T07:15       0.1246
#> 7  2011-05-24T07:45       0.1148
#> 8  2011-05-24T08:15       0.1506
#> 9  2011-05-24T08:45       0.1396
#> 10 2011-05-24T09:15       0.1558
#> # ... with 4,205 more rows
#> 
#> [[6]]
#> <Pangaea data> 10.1594/PANGAEA.811072
#> # A tibble: 3,292 × 2
#>         `Date/Time` `NLF [mm/h]`
#>               <chr>        <dbl>
#> 1  2011-06-12T10:15      -0.4887
#> 2  2011-06-12T10:45      -0.1092
#> 3  2011-06-12T11:15       0.2135
#> 4  2011-06-12T11:45       0.2696
#> 5  2011-06-12T12:15       0.1619
#> 6  2011-06-12T12:45       0.1603
#> 7  2011-06-12T13:15       0.0641
#> 8  2011-06-12T13:45      -0.0571
#> 9  2011-06-12T14:15      -0.0704
#> 10 2011-06-12T14:45      -0.0447
#> # ... with 3,282 more rows
#> 
#> [[7]]
#> <Pangaea data> 10.1594/PANGAEA.811073
#> # A tibble: 3,647 × 2
#>         `Date/Time` `NLF [mm/h]`
#>               <chr>        <dbl>
#> 1  2011-06-05T00:45      -0.3481
#> 2  2011-06-05T01:15       0.5254
#> 3  2011-06-05T01:45      -0.1642
#> 4  2011-06-05T02:15       0.1954
#> 5  2011-06-05T02:45       0.3474
#> 6  2011-06-05T03:15       0.1642
#> 7  2011-06-05T03:45       0.0578
#> 8  2011-06-05T04:15       0.0011
#> 9  2011-06-05T04:45       0.1135
#> 10 2011-06-05T05:15       0.4130
#> # ... with 3,637 more rows
#> 
#> [[8]]
#> <Pangaea data> 10.1594/PANGAEA.811074
#> # A tibble: 2,145 × 2
#>         `Date/Time` `NLF [mm/h]`
#>               <chr>        <dbl>
#> 1  2011-07-06T07:45       1.7942
#> 2  2011-07-06T08:15       1.1926
#> 3  2011-07-06T08:45       0.4628
#> 4  2011-07-06T09:15       1.3922
#> 5  2011-07-06T09:45       0.9768
#> 6  2011-07-06T10:15       0.4352
#> 7  2011-07-06T10:45       0.1580
#> 8  2011-07-06T11:15      -0.0499
#> 9  2011-07-06T11:45       0.1517
#> 10 2011-07-06T12:15      -0.0004
#> # ... with 2,135 more rows
#> 
#> [[9]]
#> <Pangaea data> 10.1594/PANGAEA.811077
#> # A tibble: 4,215 × 2
#>         `Date/Time` `NLF [mm/h]`
#>               <chr>        <dbl>
#> 1  2011-05-24T04:45      -0.4233
#> 2  2011-05-24T05:15      -0.4290
#> 3  2011-05-24T05:45      -0.5916
#> 4  2011-05-24T06:15      -0.5141
#> 5  2011-05-24T06:45      -0.6595
#> 6  2011-05-24T07:15      -0.5933
#> 7  2011-05-24T07:45      -0.4699
#> 8  2011-05-24T08:15      -0.3536
#> 9  2011-05-24T08:45      -0.4149
#> 10 2011-05-24T09:15      -0.6739
#> # ... with 4,205 more rows
#> 
#> [[10]]
#> <Pangaea data> 10.1594/PANGAEA.811078
#> # A tibble: 3,647 × 2
#>         `Date/Time` `NLF [mm/h]`
#>               <chr>        <dbl>
#> 1  2011-06-05T00:45       0.0490
#> 2  2011-06-05T01:15      -0.0669
#> 3  2011-06-05T01:45      -0.0449
#> 4  2011-06-05T02:15       0.0901
#> 5  2011-06-05T02:45       0.0373
#> 6  2011-06-05T03:15       0.0382
#> 7  2011-06-05T03:45      -0.0214
#> 8  2011-06-05T04:15      -0.0148
#> 9  2011-06-05T04:45       0.0573
#> 10 2011-06-05T05:15       0.0705
#> # ... with 3,637 more rows
#> 

pg_data(res$doi[4])

#> Downloading 1 datasets from 10.1594/PANGAEA.759113

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.759113
#> # A tibble: 20 × 29
#>            Event Latitude Longitude `Depth [m]` `Depth top [m]` `Depth bot [m]`
#>            <chr>    <dbl>     <dbl>       <dbl>           <int>           <dbl>
#> 1   Yur-8-9/87-1  50.6880  155.9038        0.25               0             0.5
#> 2   Yur-8-9/87-2  50.6884  155.9033        0.25               0             0.5
#> 3   Yur-8-9/87-3  50.6888  155.9025        0.25               0             0.5
#> 4   Yur-8-9/87-4  50.6893  155.9014        0.25               0             0.5
#> 5   Yur-8-9/87-5  50.6897  155.9007        0.25               0             0.5
#> 6   Yur-8-9/87-6  50.6900  155.9001        0.25               0             0.5
#> 7   Yur-8-9/87-7  50.6906  155.8990        0.25               0             0.5
#> 8   Yur-8-9/87-8  50.6889  155.9080        0.25               0             0.5
#> 9   Yur-8-9/87-9  50.6894  155.9074        0.25               0             0.5
#> 10 Yur-8-9/87-10  50.6897  155.9068        0.25               0             0.5
#> 11 Yur-8-9/87-11  50.6902  155.9063        0.25               0             0.5
#> 12 Yur-8-9/87-12  50.6907  155.9054        0.25               0             0.5
#> 13 Yur-8-9/87-13  50.6913  155.9045        0.25               0             0.5
#> 14 Yur-8-9/87-14  50.6917  155.9038        0.25               0             0.5
#> 15 Yur-8-9/87-15  50.6901  155.9096        0.25               0             0.5
#> 16 Yur-8-9/87-16  50.6907  155.9089        0.25               0             0.5
#> 17 Yur-8-9/87-17  50.6912  155.9084        0.25               0             0.5
#> 18 Yur-8-9/87-18  50.6917  155.9075        0.25               0             0.5
#> 19 Yur-8-9/87-19  50.6921  155.9069        0.25               0             0.5
#> 20 Yur-8-9/87-20  50.6927  155.9062        0.25               0             0.5
#> # ... with 23 more variables: pH <dbl>, `Cl [g/l]` <dbl>, `TPM [mg/l]` <dbl>,
#> #   `Fe [µg/l]` <dbl>, `Fe part [mg/l]` <int>, `Fe part [%]` <dbl>, `Fe
#> #   [%]` <dbl>, `Mn2+ [µg/l]` <dbl>, `Mn [mg/l]` <dbl>, `Mn part [%]` <dbl>,
#> #   `Mn [%]` <dbl>, `Ni [µg/l]` <dbl>, `Ni part [mg/l]` <dbl>, `Ni part
#> #   [%]` <dbl>, `Ni [%]` <dbl>, `Cu [µg/l]` <dbl>, `Cu part [mg/l]` <dbl>, `Cu
#> #   part [%]` <dbl>, `Cu [%]` <dbl>, `Zn [µg/l]` <dbl>, `Zn part [mg/l]` <dbl>,
#> #   `Zn part [%]` <dbl>, `Zn [%]` <dbl>
#>