Download data from Pangaea.

Grabs data as a dataframe or list of dataframes from a Pangaea data repository URI; see: https://www.pangaea.de/

pg_cache_clear(doi = NULL, prompt = TRUE)

pg_cache_list()

pg_data(doi, overwrite = TRUE, verbose = TRUE, ...)

Arguments

doi

DOI of Pangaeae single dataset, or of a collection of datasets. Expects either just a DOI of the form 10.1594/PANGAEA.746398, or with the URL part in front, like https://doi.pangaea.de/10.1594/PANGAEA.746398

prompt

(logical) Prompt before clearing all files in cache? No prompt used when DOIs assed in. Default: TRUE

overwrite

(logical) Ovewrite a file if one is found with the same name

verbose

(logical) print information messages. Default: TRUE

...

Curl options passed on to httr::GET()

Value

One or more items of class pangaea, each with the doi, parent doi (if many dois within a parent doi), url, citation, path, and data object. Data object depends on what kind of file it is. For tabular data, we print the first 10 columns or so; for a zip file we list the files in the zip (but leave it up to the user to dig unzip and get files from the zip file); for png files, we point the user to read the file in with png::readPNG()

Details

Data files are stored in an operating system appropriate location. Run rappdirs::user_cache_dir("pangaear") to get the storage location on your machine.

Some files/datasets require the user to be logged in. For now we just pass on these - that is, give back nothing other than metadata.

References

https://www.pangaea.de

Examples

# a single file
(res <- pg_data(doi='10.1594/PANGAEA.807580'))

#> Downloading 1 datasets from 10.1594/PANGAEA.807580

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.807580
#>   parent doi: 10.1594/PANGAEA.807580
#>   url:        https://doi.org/10.1594/PANGAEA.807580
#>   citation:   Schiebel, Ralf; Waniek, Joanna J; Bork, Matthias; Hemleben, Christoph (2001): Physical oceanography during METEOR cruise M36/6. doi:10.1594/PANGAEA.807580,In supplement to: Schiebel, R et al. (2001): Planktic foraminiferal production stimulated by chlorophyll redistribution and entrainment of nutrients. Deep Sea Research Part I: Oceanographic Research Papers, 48(3), 721-740, doi:10.1016/S0967-0637(00)00065-0
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_807580.txt
#>   data:
#> # A tibble: 32,179 × 13
#>                Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>                <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 2  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 3  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 4  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 5  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 6  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 7  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 8  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 9  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 10 M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> # ... with 32,169 more rows, and 8 more variables: `Depth water [m]` <dbl>,
#> #   `Press [dbar]` <int>, `Temp [°C]` <dbl>, Sal <dbl>, `Tpot [°C]` <dbl>,
#> #   `Sigma-theta [kg/m**3]` <dbl>, `Sigma in situ [kg/m**3]` <dbl>, `Cond
#> #   [mS/cm]` <dbl>
#> 

res[[1]]$doi

#> [1] "10.1594/PANGAEA.807580"

res[[1]]$citation

#> [1] "Schiebel, Ralf; Waniek, Joanna J; Bork, Matthias; Hemleben, Christoph (2001): Physical oceanography during METEOR cruise M36/6. doi:10.1594/PANGAEA.807580,In supplement to: Schiebel, R et al. (2001): Planktic foraminiferal production stimulated by chlorophyll redistribution and entrainment of nutrients. Deep Sea Research Part I: Oceanographic Research Papers, 48(3), 721-740, doi:10.1016/S0967-0637(00)00065-0"

res[[1]]$data

#> # A tibble: 32,179 × 13
#>                Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>                <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 2  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 3  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 4  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 5  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 6  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 7  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 8  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 9  M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> 10 M36/6-CTD-365_003 1996-10-14T12:24  48.9667  -16.4633           -4802
#> # ... with 32,169 more rows, and 8 more variables: `Depth water [m]` <dbl>,
#> #   `Press [dbar]` <int>, `Temp [°C]` <dbl>, Sal <dbl>, `Tpot [°C]` <dbl>,
#> #   `Sigma-theta [kg/m**3]` <dbl>, `Sigma in situ [kg/m**3]` <dbl>, `Cond
#> #   [mS/cm]` <dbl>

# another single file
pg_data(doi='10.1594/PANGAEA.807584')

#> Downloading 1 datasets from 10.1594/PANGAEA.807584

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.807584
#>   parent doi: 10.1594/PANGAEA.807584
#>   url:        https://doi.org/10.1594/PANGAEA.807584
#>   citation:   Schiebel, Ralf; Waniek, Joanna J; Bork, Matthias; Hemleben, Christoph (2001): Plasm bearing foraminifera counts during METEOR cruise M36/5. doi:10.1594/PANGAEA.807584,In supplement to: Schiebel, Ralf; Waniek, Joanna J; Bork, Matthias; Hemleben, Christoph (2001): Planktic foraminiferal production stimulated by chlorophyll redistribution and entrainment of nutrients. Deep Sea Research Part I: Oceanographic Research Papers, 48(3), 721-740, doi:10.1016/S0967-0637(00)00065-0
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_807584.txt
#>   data:
#> # A tibble: 698 × 53
#>            Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>            <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 2  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 3  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 4  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 5  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 6  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 7  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 8  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 9  M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> 10 M36/5_MSN1121 1996-09-10T07:46  47.1844  -19.5672           -4567
#> # ... with 688 more rows, and 48 more variables: `Depth water [m]` <int>,
#> #   `Depth top [m]` <int>, `Depth bot [m]` <int>, Size <chr>, `D. anfracta
#> #   [#]` <int>, `G. bulloides [#]` <int>, `G. calida [#]` <int>, `T. clarkei
#> #   [#]` <int>, `T. compressa [#]` <int>, `G. conglobatus [#]` <int>, `G.
#> #   crassaformis [#]` <int>, `S. dehiscens [#]` <int>, `B. digitata [#]` <int>,
#> #   `G. digitata [#]` <int>, `H. digitata [#]` <int>, `N. dutertrei [#]` <int>,
#> #   `G. falconensis [#]` <int>, `S. globigerum [#]` <int>, `G. glutinata
#> #   [#]` <int>, `G. hirsuta [#]` <int>, `T. humilis [#]` <int>, `N. incompta
#> #   [#]` <int>, `G. inflata [#]` <int>, `T. iota [#]` <int>, `G. menardii
#> #   [#]` <int>, `G. minuta [#]` <int>, `C. nitida [#]` <int>, `P.
#> #   obliquiloculata [#]` <int>, `N. pachyderma [#]` <int>, `T. parkerae
#> #   [#]` <int>, `H. pelagica [#]` <int>, `B. pumilio [#]` <int>, `T.
#> #   quinqueloba [#]` <int>, `G. ruber p [#]` <int>, `G. ruber w [#]` <int>, `G.
#> #   rubescens [#]` <int>, `G. sacculifer [#]` <int>, `G. scitula [#]` <int>,
#> #   `G. siphonifera [#]` <int>, `G. tenella [#]` <int>, `G. theyeri [#]` <int>,
#> #   `G. truncatulinoides [#]` <int>, `G. tumida [#]` <int>, `G. ungulata
#> #   [#]` <int>, `O. universa [#]` <int>, `G. uvula [#]` <int>, `G. vivans
#> #   [#]` <int>, `Foram plankt indet [#]` <int>
#> 

# Many files
(res <- pg_data(doi='10.1594/PANGAEA.761032'))

#> Downloading 2 datasets from 10.1594/PANGAEA.761032

#> Processing 2 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.761002
#>   parent doi: 10.1594/PANGAEA.761032
#>   url:        https://doi.org/10.1594/PANGAEA.761002
#>   citation:   Gordon, Arnold L; Visbeck, Martin; Huber, Bruce (2001): Physical oceanography and current velocity during Nathaniel B. Palmer cruise NBP97-05. doi:10.1594/PANGAEA.761032,Supplement to: Gordon, AL et al. (2001): Export of Weddell Sea deep and bottom water. Journal of Geophysical Research-Oceans, 106(C5), 9005-9017, doi:10.1029/2000JC000281
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_761002.txt
#>   data:
#> # A tibble: 250,201 × 9
#>          Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>          <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 2  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 3  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 4  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 5  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 6  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 7  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 8  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 9  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 10 NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> # ... with 250,191 more rows, and 4 more variables: `Depth water [m]` <int>,
#> #   `Temp [°C]` <dbl>, Sal <dbl>, `Tpot [°C]` <dbl>
#> 
#> [[2]]
#> <Pangaea data> 10.1594/PANGAEA.319391
#>   parent doi: 10.1594/PANGAEA.761032
#>   url:        https://doi.org/10.1594/PANGAEA.319391
#>   citation:   Gordon, Arnold L; Visbeck, Martin; Huber, Bruce (2001): Physical oceanography and current velocity during Nathaniel B. Palmer cruise NBP97-05. doi:10.1594/PANGAEA.761032,Supplement to: Gordon, AL et al. (2001): Export of Weddell Sea deep and bottom water. Journal of Geophysical Research-Oceans, 106(C5), 9005-9017, doi:10.1029/2000JC000281
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_319391.txt
#>   data:
#> # A tibble: 6,060 × 12
#>         `Date/Time` Latitude Longitude `Depth water [m]` `UC [cm/s]`
#>               <chr>    <dbl>     <dbl>             <int>       <dbl>
#> 1  1997-08-03T19:00 -56.3617  -50.6079                40        53.8
#> 2  1997-08-03T19:00 -56.3617  -50.6079                50        51.3
#> 3  1997-08-03T19:00 -56.3617  -50.6079                60        50.9
#> 4  1997-08-03T19:00 -56.3617  -50.6079                70        51.0
#> 5  1997-08-03T19:00 -56.3617  -50.6079                80        49.8
#> 6  1997-08-03T19:00 -56.3617  -50.6079                90        49.5
#> 7  1997-08-03T19:00 -56.3617  -50.6079               100        47.7
#> 8  1997-08-03T19:00 -56.3617  -50.6079               110        48.3
#> 9  1997-08-03T19:00 -56.3617  -50.6079               120        48.9
#> 10 1997-08-03T19:00 -56.3617  -50.6079               130        48.7
#> # ... with 6,050 more rows, and 7 more variables: `VC [cm/s]` <dbl>, `T tech
#> #   [°C]` <dbl>, `T tech std dev [±]` <dbl>, `Ship velo SU [m/s]` <dbl>, `Ship
#> #   velo SU std dev [±]` <dbl>, `Ship velo SV [m/s]` <dbl>, `Ship velo SV std
#> #   dev [±]` <dbl>
#> 

res[[1]]

#> <Pangaea data> 10.1594/PANGAEA.761002
#>   parent doi: 10.1594/PANGAEA.761032
#>   url:        https://doi.org/10.1594/PANGAEA.761002
#>   citation:   Gordon, Arnold L; Visbeck, Martin; Huber, Bruce (2001): Physical oceanography and current velocity during Nathaniel B. Palmer cruise NBP97-05. doi:10.1594/PANGAEA.761032,Supplement to: Gordon, AL et al. (2001): Export of Weddell Sea deep and bottom water. Journal of Geophysical Research-Oceans, 106(C5), 9005-9017, doi:10.1029/2000JC000281
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_761002.txt
#>   data:
#> # A tibble: 250,201 × 9
#>          Event      `Date/Time` Latitude Longitude `Elevation [m]`
#>          <chr>            <chr>    <dbl>     <dbl>           <int>
#> 1  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 2  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 3  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 4  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 5  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 6  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 7  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 8  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 9  NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> 10 NBP97-05/01 1997-08-04T12:09 -57.5007  -44.9915           -3090
#> # ... with 250,191 more rows, and 4 more variables: `Depth water [m]` <int>,
#> #   `Temp [°C]` <dbl>, Sal <dbl>, `Tpot [°C]` <dbl>

res[[2]]

#> <Pangaea data> 10.1594/PANGAEA.319391
#>   parent doi: 10.1594/PANGAEA.761032
#>   url:        https://doi.org/10.1594/PANGAEA.319391
#>   citation:   Gordon, Arnold L; Visbeck, Martin; Huber, Bruce (2001): Physical oceanography and current velocity during Nathaniel B. Palmer cruise NBP97-05. doi:10.1594/PANGAEA.761032,Supplement to: Gordon, AL et al. (2001): Export of Weddell Sea deep and bottom water. Journal of Geophysical Research-Oceans, 106(C5), 9005-9017, doi:10.1029/2000JC000281
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_319391.txt
#>   data:
#> # A tibble: 6,060 × 12
#>         `Date/Time` Latitude Longitude `Depth water [m]` `UC [cm/s]`
#>               <chr>    <dbl>     <dbl>             <int>       <dbl>
#> 1  1997-08-03T19:00 -56.3617  -50.6079                40        53.8
#> 2  1997-08-03T19:00 -56.3617  -50.6079                50        51.3
#> 3  1997-08-03T19:00 -56.3617  -50.6079                60        50.9
#> 4  1997-08-03T19:00 -56.3617  -50.6079                70        51.0
#> 5  1997-08-03T19:00 -56.3617  -50.6079                80        49.8
#> 6  1997-08-03T19:00 -56.3617  -50.6079                90        49.5
#> 7  1997-08-03T19:00 -56.3617  -50.6079               100        47.7
#> 8  1997-08-03T19:00 -56.3617  -50.6079               110        48.3
#> 9  1997-08-03T19:00 -56.3617  -50.6079               120        48.9
#> 10 1997-08-03T19:00 -56.3617  -50.6079               130        48.7
#> # ... with 6,050 more rows, and 7 more variables: `VC [cm/s]` <dbl>, `T tech
#> #   [°C]` <dbl>, `T tech std dev [±]` <dbl>, `Ship velo SU [m/s]` <dbl>, `Ship
#> #   velo SU std dev [±]` <dbl>, `Ship velo SV [m/s]` <dbl>, `Ship velo SV std
#> #   dev [±]` <dbl>

# Manipulating the cache
## list files in the cache
pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_807580.txt"
#> [13] "10_1594_PANGAEA_807584.txt" "10_1594_PANGAEA_825423.txt"
#> [15] "10_1594_PANGAEA_825424.txt" "10_1594_PANGAEA_825425.txt"
#> [17] "10_1594_PANGAEA_825426.txt" "10_1594_PANGAEA_825427.png"
#> [19] "10_1594_PANGAEA_860500.zip"

## clear all data
# pg_cache_clear()
pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_807580.txt"
#> [13] "10_1594_PANGAEA_807584.txt" "10_1594_PANGAEA_825423.txt"
#> [15] "10_1594_PANGAEA_825424.txt" "10_1594_PANGAEA_825425.txt"
#> [17] "10_1594_PANGAEA_825426.txt" "10_1594_PANGAEA_825427.png"
#> [19] "10_1594_PANGAEA_860500.zip"

## clear a single dataset by DOI
pg_data(doi='10.1594/PANGAEA.812093')

#> Downloading 1 datasets from 10.1594/PANGAEA.812093

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.812093
#>   parent doi: 10.1594/PANGAEA.812093
#>   url:        https://doi.org/10.1594/PANGAEA.812093
#>   citation:   Simonyan, Anna V; Dultz, Stefan; Behrens, Harald (2012): (Table 2) Conditions and results of diffusion experiments of East Pacific Ocean basalt samples. doi:10.1594/PANGAEA.812093,In supplement to: Simonyan, AV et al. (2012): Diffusive transport of water in porous fresh to altered mid-ocean ridge basalts. Chemical Geology, 306-307, 63-77, doi:10.1016/j.chemgeo.2012.02.017
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_812093.txt
#>   data:
#> # A tibble: 36 × 18
#>       Event `Sample label`                              `Sample comment`
#>       <chr>          <chr>                                         <chr>
#> 1  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 2  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 3  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 4  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 5  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 6  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 7  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 8  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 9  169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> 10 169-856H 169-856H-55R-1 complete altered, thickness of basalt =200 µm
#> # ... with 26 more rows, and 15 more variables: `Depth [m]` <dbl>, `Distance
#> #   [µm]` <int>, `T tech [°C]` <dbl>, Run <chr>, `- (initial concentration of
#> #   diff...)` <dbl>, `- (initial concentration of diff...)` <dbl>, `Poros frac
#> #   (#1, from fit)` <dbl>, `Poros frac (#2, from fit)` <dbl>, `Diff coeff
#> #   [10**-6 cm**2/s]` <dbl>, `Diff coeff std dev [±] (effective from
#> #   fit)` <dbl>, `Diff coeff std dev [±] (effective total)` <dbl>, X <dbl>, `X
#> #   (average)` <int>, F <int>, `- (m)` <dbl>
#> 

pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_807580.txt"
#> [13] "10_1594_PANGAEA_807584.txt" "10_1594_PANGAEA_812093.txt"
#> [15] "10_1594_PANGAEA_825423.txt" "10_1594_PANGAEA_825424.txt"
#> [17] "10_1594_PANGAEA_825425.txt" "10_1594_PANGAEA_825426.txt"
#> [19] "10_1594_PANGAEA_825427.png" "10_1594_PANGAEA_860500.zip"

pg_cache_clear(doi='10.1594/PANGAEA.812093')
pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_807580.txt"
#> [13] "10_1594_PANGAEA_807584.txt" "10_1594_PANGAEA_825423.txt"
#> [15] "10_1594_PANGAEA_825424.txt" "10_1594_PANGAEA_825425.txt"
#> [17] "10_1594_PANGAEA_825426.txt" "10_1594_PANGAEA_825427.png"
#> [19] "10_1594_PANGAEA_860500.zip"

## clear more than 1 dataset by DOI
lapply(c('10.1594/PANGAEA.746398','10.1594/PANGAEA.746400'), pg_data)

#> Downloading 1 datasets from 10.1594/PANGAEA.746398

#> Processing 1 files

#> Downloading 1 datasets from 10.1594/PANGAEA.746400

#> Processing 1 files

#> [[1]]
#> [[1]][[1]]
#> <Pangaea data> 10.1594/PANGAEA.746398
#>   parent doi: 10.1594/PANGAEA.746398
#>   url:        https://doi.org/10.1594/PANGAEA.746398
#>   citation:   Sval'nov, Vyacheslav N; Alekseeva, Tatyana N (2005): Grain size composition of sediments from Core PS-279. P.P. Shirshov Institute of Oceanology, Russian Academy of Sciences, Moscow, doi:10.1594/PANGAEA.746398,In supplement to: Sval'nov, VN; Alekseeva, TN (2005): Grain Size Composition of Bottom Sediments of the World Ocean. Nauka Publ. (Moscow): in Russian, 297 pp
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_746398.txt
#>   data:
#> # A tibble: 4 × 13
#>   `Depth [m]` `Depth top [m]` `Depth bot [m]`         Sediment `1-0.5 mm [%]`
#>         <dbl>           <dbl>           <dbl>            <chr>          <dbl>
#> 1       0.025            0.00            0.05 coccolithic ooze           0.44
#> 2       0.160            0.15            0.17 coccolithic ooze           5.05
#> 3       0.175            0.35            0.00 coccolithic ooze           2.85
#> 4       0.205            0.17            0.24 coccolithic ooze           8.12
#> # ... with 8 more variables: `500-250 µm [%]` <dbl>, `250-100 µm [%]` <dbl>,
#> #   `100-50 µm [%]` <dbl>, `50-10 µm [%]` <dbl>, `10-5 µm [%]` <dbl>, `5-1 µm
#> #   [%]` <dbl>, `<1 µm [%]` <dbl>, `Sum [%]` <dbl>
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> <Pangaea data> 10.1594/PANGAEA.746400
#>   parent doi: 10.1594/PANGAEA.746400
#>   url:        https://doi.org/10.1594/PANGAEA.746400
#>   citation:   Sval'nov, Vyacheslav N; Alekseeva, Tatyana N (2005): Grain size composition of sediments from Core PS-281. P.P. Shirshov Institute of Oceanology, Russian Academy of Sciences, Moscow, doi:10.1594/PANGAEA.746400,In supplement to: Sval'nov, VN; Alekseeva, TN (2005): Grain Size Composition of Bottom Sediments of the World Ocean. Nauka Publ. (Moscow): in Russian, 297 pp
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_746400.txt
#>   data:
#> # A tibble: 11 × 13
#>    `Depth [m]` `Depth top [m]` `Depth bot [m]`           Sediment
#>          <dbl>           <dbl>           <dbl>              <chr>
#> 1        0.025            0.00            0.05 foraminiferal ooze
#> 2        0.025            0.00            0.05 foraminiferal ooze
#> 3        0.375            0.35            0.40 foraminiferal ooze
#> 4        0.375            0.35            0.40 foraminiferal ooze
#> 5        0.475            0.45            0.50   coccolithic ooze
#> 6        1.025            1.00            1.05 foraminiferal ooze
#> 7        1.025            1.00            1.05 foraminiferal ooze
#> 8        1.725            1.70            1.75 foraminiferal ooze
#> 9        1.725            1.70            1.75 foraminiferal ooze
#> 10       2.475            2.45            2.50   coccolithic ooze
#> 11       3.105            3.08            3.13   coccolithic ooze
#> # ... with 9 more variables: `1-0.5 mm [%]` <dbl>, `500-250 µm [%]` <dbl>,
#> #   `250-100 µm [%]` <dbl>, `100-50 µm [%]` <dbl>, `50-10 µm [%]` <dbl>, `10-5
#> #   µm [%]` <dbl>, `5-1 µm [%]` <dbl>, `<1 µm [%]` <dbl>, `Sum [%]` <dbl>
#> 
#> 

pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_746398.txt"
#>  [3] "10_1594_PANGAEA_746400.txt" "10_1594_PANGAEA_761002.txt"
#>  [5] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [7] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [9] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#> [11] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [13] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_807580.txt"
#> [15] "10_1594_PANGAEA_807584.txt" "10_1594_PANGAEA_825423.txt"
#> [17] "10_1594_PANGAEA_825424.txt" "10_1594_PANGAEA_825425.txt"
#> [19] "10_1594_PANGAEA_825426.txt" "10_1594_PANGAEA_825427.png"
#> [21] "10_1594_PANGAEA_860500.zip"

pg_cache_clear(doi=c('10.1594/PANGAEA.746398','10.1594/PANGAEA.746400'))
pg_cache_list()

#>  [1] "10_1594_PANGAEA_319391.txt" "10_1594_PANGAEA_761002.txt"
#>  [3] "10_1594_PANGAEA_77397.txt"  "10_1594_PANGAEA_77398.txt" 
#>  [5] "10_1594_PANGAEA_77399.txt"  "10_1594_PANGAEA_77400.txt" 
#>  [7] "10_1594_PANGAEA_77401.txt"  "10_1594_PANGAEA_77402.txt" 
#>  [9] "10_1594_PANGAEA_77403.txt"  "10_1594_PANGAEA_77404.txt" 
#> [11] "10_1594_PANGAEA_77406.txt"  "10_1594_PANGAEA_807580.txt"
#> [13] "10_1594_PANGAEA_807584.txt" "10_1594_PANGAEA_825423.txt"
#> [15] "10_1594_PANGAEA_825424.txt" "10_1594_PANGAEA_825425.txt"
#> [17] "10_1594_PANGAEA_825426.txt" "10_1594_PANGAEA_825427.png"
#> [19] "10_1594_PANGAEA_860500.zip"

# search for datasets, then pass in DOIs
(searchres <- pg_search(query = 'birds', count = 20))

#> # A tibble: 20 × 6
#>        score                    doi  size size_measure
#> *      <dbl>                  <chr> <dbl>        <chr>
#> 1  105.46468 10.1594/PANGAEA.743406     2     datasets
#> 2   69.22585 10.1594/PANGAEA.788547     3     datasets
#> 3   68.81578 10.1594/PANGAEA.839226    74  data points
#> 4   64.36520 10.1594/PANGAEA.808549    80  data points
#> 5   63.67343 10.1594/PANGAEA.633712    16     datasets
#> 6   60.86255 10.1594/PANGAEA.817547     3     datasets
#> 7   60.61428 10.1594/PANGAEA.872356   220  data points
#> 8   59.76067 10.1594/PANGAEA.837408    92  data points
#> 9   57.46819 10.1594/PANGAEA.786530   363  data points
#> 10  56.91473 10.1594/PANGAEA.841257 38990  data points
#> 11  55.95452 10.1594/PANGAEA.816022     4     datasets
#> 12  52.03966 10.1594/PANGAEA.841258    75  data points
#> 13  50.92400 10.1594/PANGAEA.808873    39  data points
#> 14  50.85811 10.1594/PANGAEA.761424     2     datasets
#> 15  50.85655 10.1594/PANGAEA.841264     3     datasets
#> 16  48.45797 10.1594/PANGAEA.633713    30     datasets
#> 17  46.88770 10.1594/PANGAEA.815269    20  data points
#> 18  44.58861 10.1594/PANGAEA.841252 25236  data points
#> 19  42.99368 10.1594/PANGAEA.819304   186  data points
#> 20  41.71456 10.1594/PANGAEA.815229     3     datasets
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_data(searchres$doi[1])

#> Downloading 2 datasets from 10.1594/PANGAEA.743406

#> Processing 2 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.743401
#>   parent doi: 10.1594/PANGAEA.743406
#>   url:        https://doi.org/10.1594/PANGAEA.743401
#>   citation:   Novatti, Ricardo (1963): Bird observations in the Weddell Sea. doi:10.1594/PANGAEA.743406,Supplement to: Novatti, R (1963): Pelagic distribution of birds at the Weddell Sea. Polarforschung, 33(1/2), 207-213, hdl:10013/epic.29260.d001
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_743401.txt
#>   data:
#> # A tibble: 45 × 23
#>    `Date/Time` Latitude Longitude `P. atriceps [#]` `S. vittata [#]`
#>          <chr>    <dbl>     <dbl>             <int>            <chr>
#> 1   1955-12-06  -63.983   -55.067                 0                0
#> 2   1955-12-07  -61.233   -46.250                 0                0
#> 3   1955-12-08  -60.750   -44.717                 0               ?1
#> 4   1955-12-09  -60.750   -44.717                 2               ?1
#> 5   1955-12-10  -60.750   -44.717                22                2
#> 6   1955-12-11  -60.750   -44.717                 7                1
#> 7   1955-12-12  -60.133   -37.267                 0                0
#> 8   1955-12-13  -59.500   -28.133                 0                0
#> 9   1955-12-14  -59.450   -27.250                 0                0
#> 10  1955-12-15  -64.317   -36.467                 0                0
#> # ... with 35 more rows, and 18 more variables: `L. dominicanus [#]` <chr>, `C.
#> #   skua [#]` <chr>, `C. alba [#]` <int>, `O. oceanicus [#]` <chr>, `H.
#> #   caerulea [#]` <chr>, `P. desolata [#]` <chr>, `P. nivea [#]` <chr>, `D.
#> #   capensis [#]` <chr>, `T. antarctica [#]` <chr>, `F. glacialoides
#> #   [#]` <chr>, `M. giganteus [#]` <int>, `D. exulans exulans [#]` <int>, `E.
#> #   chrysolophus [#]` <int>, `P. papua [#]` <chr>, `P. antarctica [#]` <chr>,
#> #   `P. adeliae [#]` <chr>, `A. forsteri [#]` <chr>, Comment <chr>
#> 
#> [[2]]
#> <Pangaea data> 10.1594/PANGAEA.743405
#>   parent doi: 10.1594/PANGAEA.743406
#>   url:        https://doi.org/10.1594/PANGAEA.743405
#>   citation:   Novatti, Ricardo (1963): Bird observations in the Weddell Sea. doi:10.1594/PANGAEA.743406,Supplement to: Novatti, R (1963): Pelagic distribution of birds at the Weddell Sea. Polarforschung, 33(1/2), 207-213, hdl:10013/epic.29260.d001
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_743405.txt
#>   data:
#> # A tibble: 24 × 22
#>    `Date/Time` Latitude Longitude `C. skua [#]` `C. alba [#]`
#>          <chr>    <dbl>     <dbl>         <int>         <int>
#> 1   1959-12-22  -63.833   -56.183             0             0
#> 2   1959-12-23  -62.233   -51.167             0             1
#> 3   1959-12-24  -61.200   -45.217             0             0
#> 4   1959-12-25  -60.183   -38.317             0             0
#> 5   1959-12-26  -60.200   -31.333             0             1
#> 6   1959-12-27  -61.333   -23.067             0             0
#> 7   1959-12-28  -65.200   -21.833             0             0
#> 8   1959-12-29  -66.967   -18.333             0             0
#> 9   1959-12-30  -69.333   -16.150             0             0
#> 10  1959-12-31  -68.667   -14.100             0             0
#> # ... with 14 more rows, and 17 more variables: `O. oceanicus [#]` <int>, `H.
#> #   caerulea [#]` <int>, `P. macro. macroptera [#]` <int>, `P. desolata
#> #   [#]` <int>, `P. nivea [#]` <int>, `T. antarctica [#]` <int>, `D. capensis
#> #   [#]` <int>, `F. glacialoides [#]` <int>, `M. giganteus [#]` <int>, `D.
#> #   exulans exulans [#]` <int>, `D. chrysostoma [#]` <int>, `D. melanophris
#> #   [#]` <int>, `P. palpebrata [#]` <int>, `P. fusca [#]` <int>, `P. antarctica
#> #   [#]` <int>, `P. adeliae [#]` <int>, `A. forsteri [#]` <int>
#> 

pg_data(searchres$doi[2])

#> Downloading 3 datasets from 10.1594/PANGAEA.788547

#> Warning: Log in required, skipping file download

#> Warning: Log in required, skipping file download

#> Warning: Log in required, skipping file download

#> Processing 3 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.788545
#>   parent doi: 10.1594/PANGAEA.788547
#>   url:        https://doi.org/10.1594/PANGAEA.788545
#>   citation:   Muñoz, Juan C; Aerts, Raf; Thijs, Koen W; Stevenson, Pablo R; Muys, Bart; Sekercioglu, Cagan H (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. doi:10.1594/PANGAEA.788547,Supplement to: Muñoz, JC et al. (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. Agriculture Ecosystems & Environment, 173, 13-19, doi:10.1016/j.agee.2013.04.006
#>   path:       NA
#>   data:
#> [1] NA
#> 
#> [[2]]
#> <Pangaea data> 10.1594/PANGAEA.788544
#>   parent doi: 10.1594/PANGAEA.788547
#>   url:        https://doi.org/10.1594/PANGAEA.788544
#>   citation:   Muñoz, Juan C; Aerts, Raf; Thijs, Koen W; Stevenson, Pablo R; Muys, Bart; Sekercioglu, Cagan H (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. doi:10.1594/PANGAEA.788547,Supplement to: Muñoz, JC et al. (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. Agriculture Ecosystems & Environment, 173, 13-19, doi:10.1016/j.agee.2013.04.006
#>   path:       NA
#>   data:
#> [1] NA
#> 
#> [[3]]
#> <Pangaea data> 10.1594/PANGAEA.788546
#>   parent doi: 10.1594/PANGAEA.788547
#>   url:        https://doi.org/10.1594/PANGAEA.788546
#>   citation:   Muñoz, Juan C; Aerts, Raf; Thijs, Koen W; Stevenson, Pablo R; Muys, Bart; Sekercioglu, Cagan H (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. doi:10.1594/PANGAEA.788547,Supplement to: Muñoz, JC et al. (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. Agriculture Ecosystems & Environment, 173, 13-19, doi:10.1016/j.agee.2013.04.006
#>   path:       NA
#>   data:
#> [1] NA
#> 

pg_data(searchres$doi[3])

#> Downloading 1 datasets from 10.1594/PANGAEA.839226

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.839226
#>   parent doi: 10.1594/PANGAEA.839226
#>   url:        https://doi.org/10.1594/PANGAEA.839226
#>   citation:   Thierry, Anne-Mathilde; Massemin, Sylvie; Handrich, Yves; Raclot, Thierry (2013): (Table 1) Breeding parameters for 60 male Adélie penguins (Pygoscelis adeliae) at station Dumont d'Urville (Antarctica) depending on treatment (control vs. CORT). doi:10.1594/PANGAEA.839226,Supplement to: Thierry, A-M et al. (2013): Elevated corticosterone levels and severe weather conditions decrease parental investment of incubating Adélie penguins. Hormones and Behavior, 63(3), 475-483, doi:10.1016/j.yhbeh.2012.12.011
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_839226.txt
#>   data:
#> # A tibble: 4 × 23
#>    Coverage      `Method comm`   Treat `P. adeliae [#]`
#>       <chr>              <chr>   <chr>            <int>
#> 1 2007-2010 Without dummy eggs Control               23
#> 2 2007-2010 Without dummy eggs    CORT               23
#> 3 2007-2010    With dummy eggs Control                5
#> 4 2007-2010    With dummy eggs    CORT                9
#> # ... with 19 more variables: `Treat (date of implantation)` <chr>, `Std dev
#> #   [±] (implantation date)` <dbl>, `P. adeliae m [kg]` <dbl>, `P. adeliae std
#> #   dev [±] (male body mass)` <dbl>, `P. adeliae egg laying` <chr>, `P. adeliae
#> #   std dev [±] (egg laying date)` <dbl>, `Treat (date of female
#> #   departure)` <chr>, `Std dev [±] (female departure date)` <dbl>,
#> #   `Eggs/clutch [#]` <dbl>, `Eggs/clutch std e [±]` <dbl>, `Treat (date start
#> #   of incubation beha...)` <chr>, `Std dev [±] (start of behaviour
#> #   recording)` <dbl>, `Treat (date end of incubation behavi...)` <chr>, `Std
#> #   dev [±] (end of behaviour recording)` <dbl>, `Treat (date of male
#> #   departure)` <chr>, `Std dev [±] (male departure date)` <dbl>, `P. adeliae
#> #   hatching` <chr>, `P. adeliae std dev [±] (hatching date)` <dbl>,
#> #   Comment <chr>
#> 

pg_data(searchres$doi[4])

#> Downloading 1 datasets from 10.1594/PANGAEA.808549

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.808549
#>   parent doi: 10.1594/PANGAEA.808549
#>   url:        https://doi.org/10.1594/PANGAEA.808549
#>   citation:   Cottin, Manuelle; Kato, Akiko; Thierry, Anne-Mathilde; Le Maho, Yvon; Raclot, Thierry; Ropert-Coudert, Yan (2011): (Table 1) Diving behaviour of male Adélie penguins (Pygoscelis adeliae) with and without corticosterone implants at Pointe Géologie. doi:10.1594/PANGAEA.808549,Supplement to: Cottin, M et al. (2011): Does corticosterone affect diving behaviour of male Adelie Penguins? A preliminary experimental study. ORNITHOLOGICAL SCIENCE, 10(1), 3-11, doi:10.2326/osj.10.3
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_808549.txt
#>   data:
#> # A tibble: 4 × 20
#>     Treat        ID `Duration [days] (of recording)`
#>     <chr>     <chr>                            <dbl>
#> 1 Control Control 1                             15.7
#> 2 Control Control 2                             15.2
#> 3    CORT    CORT 3                             21.3
#> 4    CORT    CORT 4                             16.7
#> # ... with 17 more variables: `Duration [days] (per trip, mean)` <dbl>, `Std
#> #   dev [±] (mean trip duration)` <dbl>, `Dives [#]` <int>, `Std dev [±] (no.
#> #   dives/trip)` <int>, `Time [s]` <int>, `No (of bouts per trip (bout =
#> #   sho...)` <int>, `Std dev [±] (no. bouts/trip)` <int>, `No (of dives per
#> #   bout (bout = sho...)` <int>, `Std dev [±] (no.dives/bout)` <int>,
#> #   `Proportion of time` <dbl>, `Proportion of time std e [±]` <dbl>, `P.
#> #   adeliae dive max [m] (shallow dives)` <int>, `P. adeliae std dev [±] (dive
#> #   depth max shallow)` <int>, `P. adeliae dive max [m] (deep dives)` <int>,
#> #   `P. adeliae std dev [±] (dive depth max deep)` <int>, `No (of undulations
#> #   per dive)` <dbl>, `Std dev [±] (no. of undulations/dive)` <dbl>
#> 

pg_data(searchres$doi[7])

#> Downloading 1 datasets from 10.1594/PANGAEA.872356

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.872356
#>   parent doi: 10.1594/PANGAEA.872356
#>   url:        https://doi.org/10.1594/PANGAEA.872356
#>   citation:   Georgios, Goras; Chrysoula, Tananaki; Vasileios, Kanelis; Christakis, Tofaris; Epaminontas, Giannouris; Nikolia, Argena; Sofia, Gounari; Andreas, Thrasyvoulou (2017): The conflict of survival of Bee-eaters and of honey bees. Dataset #872356 (DOI registration in progress),Supplement to: Georgios, G et al. (submitted): The conflict of survival of Bee-eaters and of honey bees in Cyprus.
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_872356.txt
#>   data:
#> # A tibble: 44 × 5
#>                                                                       Content
#>                                                                         <chr>
#> 1                                 1 Two apiaries  in experimental area, video
#> 2                                                2 Experimental Apiary, video
#> 3                      3 Hanging shiny objects to chase the bee-eaters, video
#> 4                                 4 Bird sounds from Birdchaser device, video
#> 5                                     5 Scaring voices from bee eaters, video
#> 6  6 Bee-eaters form a ring over the apiary 200-300 meters from apiary, video
#> 7                               7 Restriction of bees into their hives, video
#> 8                                  8 The complete cease of bee flights, video
#> 9                  9 No flights of bees with the present of bee-eaters, video
#> 10                                10 Flight of bees without bee-eaters, video
#> # ... with 34 more rows, and 4 more variables: `File name` <chr>, `File
#> #   format` <chr>, `File size [kByte]` <dbl>, `URL file` <chr>
#> 

# png file
pg_data(doi = "10.1594/PANGAEA.825428")

#> Downloading 5 datasets from 10.1594/PANGAEA.825428

#> Processing 5 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.825427
#>   parent doi: 10.1594/PANGAEA.825428
#>   url:        https://doi.org/10.1594/PANGAEA.825427
#>   citation:   Kharanzhevskaya, Yulia Alexsandrovna (2014): Mire water balance within the River Klyuch basin, Western Siberia. doi:10.1594/PANGAEA.825428,Supplement to: Kharanzhevskaya, YA (2014): Estimation of peculiarities and long-term dynamics of mire water balance in Western Siberia. Siberian Research Institute of Agriculture and Peat, 13 pp
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_825427.png
#>   data:
#> [1] "png; read with png::readPNG()"
#> 
#> [[2]]
#> <Pangaea data> 10.1594/PANGAEA.825423
#>   parent doi: 10.1594/PANGAEA.825428
#>   url:        https://doi.org/10.1594/PANGAEA.825423
#>   citation:   Kharanzhevskaya, Yulia Alexsandrovna (2014): Mire water balance within the River Klyuch basin, Western Siberia. doi:10.1594/PANGAEA.825428,Supplement to: Kharanzhevskaya, YA (2014): Estimation of peculiarities and long-term dynamics of mire water balance in Western Siberia. Siberian Research Institute of Agriculture and Peat, 13 pp
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_825423.txt
#>   data:
#> # A tibble: 191 × 3
#>    `Date/Time` `Water level [m] (of Klyuch River given in mete...)`
#>          <chr>                                                <dbl>
#> 1   2007-04-24                                              102.271
#> 2   2007-04-25                                              102.021
#> 3   2007-04-26                                              101.891
#> 4   2007-04-27                                              101.841
#> 5   2007-04-28                                              101.771
#> 6   2007-04-29                                              101.641
#> 7   2007-04-30                                              101.531
#> 8   2007-05-01                                              101.451
#> 9   2007-05-02                                              101.381
#> 10  2007-05-03                                              101.351
#> # ... with 181 more rows, and 1 more variables: `Water level [m] (of Vasyugan
#> #   mire given in met...)` <dbl>
#> 
#> [[3]]
#> <Pangaea data> 10.1594/PANGAEA.825424
#>   parent doi: 10.1594/PANGAEA.825428
#>   url:        https://doi.org/10.1594/PANGAEA.825424
#>   citation:   Kharanzhevskaya, Yulia Alexsandrovna (2014): Mire water balance within the River Klyuch basin, Western Siberia. doi:10.1594/PANGAEA.825428,Supplement to: Kharanzhevskaya, YA (2014): Estimation of peculiarities and long-term dynamics of mire water balance in Western Siberia. Siberian Research Institute of Agriculture and Peat, 13 pp
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_825424.txt
#>   data:
#> # A tibble: 36 × 2
#>    `Date/Time` `Runoff [l/km**2/s]`
#>          <chr>                <dbl>
#> 1      2006-05                 2.49
#> 2      2006-06                 0.41
#> 3      2006-07                 0.16
#> 4      2006-08                 0.31
#> 5      2006-09                 0.17
#> 6      2006-10                 0.69
#> 7      2007-04                 5.25
#> 8      2007-05                 3.21
#> 9      2007-06                 4.96
#> 10     2007-07                 3.19
#> # ... with 26 more rows
#> 
#> [[4]]
#> <Pangaea data> 10.1594/PANGAEA.825425
#>   parent doi: 10.1594/PANGAEA.825428
#>   url:        https://doi.org/10.1594/PANGAEA.825425
#>   citation:   Kharanzhevskaya, Yulia Alexsandrovna (2014): Mire water balance within the River Klyuch basin, Western Siberia. doi:10.1594/PANGAEA.825428,Supplement to: Kharanzhevskaya, YA (2014): Estimation of peculiarities and long-term dynamics of mire water balance in Western Siberia. Siberian Research Institute of Agriculture and Peat, 13 pp
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_825425.txt
#>   data:
#> # A tibble: 60 × 2
#>    `Date/Time` `Water level [m]`
#>          <chr>             <dbl>
#> 1      1998-05           113.611
#> 2      1998-06           113.551
#> 3      1998-07           113.541
#> 4      1998-08           113.501
#> 5      1998-09           113.497
#> 6      1999-05           113.601
#> 7      1999-06           113.551
#> 8      1999-07           113.481
#> 9      1999-08           113.391
#> 10     1999-09           113.400
#> # ... with 50 more rows
#> 
#> [[5]]
#> <Pangaea data> 10.1594/PANGAEA.825426
#>   parent doi: 10.1594/PANGAEA.825428
#>   url:        https://doi.org/10.1594/PANGAEA.825426
#>   citation:   Kharanzhevskaya, Yulia Alexsandrovna (2014): Mire water balance within the River Klyuch basin, Western Siberia. doi:10.1594/PANGAEA.825428,Supplement to: Kharanzhevskaya, YA (2014): Estimation of peculiarities and long-term dynamics of mire water balance in Western Siberia. Siberian Research Institute of Agriculture and Peat, 13 pp
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_825426.txt
#>   data:
#> # A tibble: 41 × 2
#>    `Date/Time` `Evaporation [mm/a]`
#>          <chr>                <dbl>
#> 1      1970-10                  0.0
#> 2      1971-10                 36.8
#> 3      1972-10                 37.8
#> 4      1973-10                 53.6
#> 5      1974-10                 34.1
#> 6      1975-10                  0.0
#> 7      1976-10                  0.0
#> 8      1977-10                  0.0
#> 9      1978-10                 32.8
#> 10     1979-10                 94.8
#> # ... with 31 more rows
#> 

# zip file
pg_data(doi = "10.1594/PANGAEA.860500")

#> Downloading 1 datasets from 10.1594/PANGAEA.860500

#> Processing 1 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.860500
#>   parent doi: 10.1594/PANGAEA.860500
#>   url:        https://doi.org/10.1594/PANGAEA.860500
#>   citation:   Bartsch, Annett; Pathe, Carsten; Wagner, Wolfgang; Scipal, Klaus (2008): Permanent open water surfaces in central Siberia from ENVISAT ASAR wide swath data 2003/2004, link to shapefile. doi:10.1594/PANGAEA.860500,Supplement to: Bartsch, A et al. (2008): Detection of permanent open water surfaces in central Siberia with ENVISAT ASAR wide swath data with special emphasis on the estimation of methane fluxes from tundra wetlands. Hydrology Research, 39(2), 89-100, doi:10.2166/nh.2008.041
#>   path:       /Users/sacmac/Library/Caches/pangaear/10_1594_PANGAEA_860500.zip
#>   data:
#>                                                              Name  Length
#> 1                        open_water_central_siberia_ASAR_WS_2003/       0
#> 2 open_water_central_siberia_ASAR_WS_2003/complete_water_2003.dbf 1513810
#> 3 open_water_central_siberia_ASAR_WS_2003/complete_water_2003.prj     435
#> 4 open_water_central_siberia_ASAR_WS_2003/complete_water_2003.shp 8867724
#> 5 open_water_central_siberia_ASAR_WS_2003/complete_water_2003.shx  302836
#>                  Date
#> 1 2016-05-17 10:11:00
#> 2 2004-08-18 09:55:00
#> 3 2016-05-14 19:03:00
#> 4 2004-08-18 09:55:00
#> 5 2004-08-18 09:55:00
#> 

# login required
## we skip file download
pg_data("10.1594/PANGAEA.788547")

#> Downloading 3 datasets from 10.1594/PANGAEA.788547

#> Warning: Log in required, skipping file download

#> Warning: Log in required, skipping file download

#> Warning: Log in required, skipping file download

#> Processing 3 files

#> [[1]]
#> <Pangaea data> 10.1594/PANGAEA.788545
#>   parent doi: 10.1594/PANGAEA.788547
#>   url:        https://doi.org/10.1594/PANGAEA.788545
#>   citation:   Muñoz, Juan C; Aerts, Raf; Thijs, Koen W; Stevenson, Pablo R; Muys, Bart; Sekercioglu, Cagan H (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. doi:10.1594/PANGAEA.788547,Supplement to: Muñoz, JC et al. (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. Agriculture Ecosystems & Environment, 173, 13-19, doi:10.1016/j.agee.2013.04.006
#>   path:       NA
#>   data:
#> [1] NA
#> 
#> [[2]]
#> <Pangaea data> 10.1594/PANGAEA.788544
#>   parent doi: 10.1594/PANGAEA.788547
#>   url:        https://doi.org/10.1594/PANGAEA.788544
#>   citation:   Muñoz, Juan C; Aerts, Raf; Thijs, Koen W; Stevenson, Pablo R; Muys, Bart; Sekercioglu, Cagan H (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. doi:10.1594/PANGAEA.788547,Supplement to: Muñoz, JC et al. (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. Agriculture Ecosystems & Environment, 173, 13-19, doi:10.1016/j.agee.2013.04.006
#>   path:       NA
#>   data:
#> [1] NA
#> 
#> [[3]]
#> <Pangaea data> 10.1594/PANGAEA.788546
#>   parent doi: 10.1594/PANGAEA.788547
#>   url:        https://doi.org/10.1594/PANGAEA.788546
#>   citation:   Muñoz, Juan C; Aerts, Raf; Thijs, Koen W; Stevenson, Pablo R; Muys, Bart; Sekercioglu, Cagan H (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. doi:10.1594/PANGAEA.788547,Supplement to: Muñoz, JC et al. (2013): Contribution of woody habitat islands to the conservation of birds and their potential ecosystem services in an extensive Colombian rangeland. Agriculture Ecosystems & Environment, 173, 13-19, doi:10.1016/j.agee.2013.04.006
#>   path:       NA
#>   data:
#> [1] NA
#>