List identifiers of the Pangaea repository

List identifiers of the Pangaea repository

pg_list_identifiers(prefix = "oai_dc", from = NULL, until = NULL,
  set = NULL, token = NULL, as = "df", ...)

Arguments

prefix

A character string to specify the metadata format in OAI-PMH requests issued to the repository. The default (oai_dc) corresponds to the mandatory OAI unqualified Dublin Core metadata schema.

from

Character string giving datestamp to be used as lower bound for datestamp-based selective harvesting (i.e., only harvest records with datestamps in the given range). Dates and times must be encoded using ISO 8601. The trailing Z must be used when including time. OAI-PMH implies UTC for data/time specifications.

until

Character string giving a datestamp to be used as an upper bound, for datestamp-based selective harvesting (i.e., only harvest records with datestamps in the given range).

set

A character string giving a set to be used for selective harvesting (i.e., only harvest records in the given set).

token

(character) a token previously provided by the server to resume a request where it last left off. 50 is max number of records returned. We will loop for you internally to get all the records you asked for.

as

(character) What to return. One of "df" (for data.frame; default), "list", or "raw" (raw text)

...

Curl debugging options passed on to httr::GET()

Value

XML character string, data.frame, or list, depending on what requested with the as parameter

References

https://www.openarchives.org/pmh/

See also

wraps oai::list_identifiers()

Other oai methods: pg_get_record, pg_identify, pg_list_metadata_formats, pg_list_records, pg_list_sets

Examples

pg_list_identifiers(
  from = paste0(Sys.Date() - 2, "T00:00:00Z"),
  until = paste0(Sys.Date() - 1, "T18:00:00Z")
)

#> # A tibble: 19,237 × 7
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1   oai:pangaea.de:doi:10.1594/PANGAEA.52692 2017-03-05T14:44:38Z citable
#> 2   oai:pangaea.de:doi:10.1594/PANGAEA.53178 2017-03-05T14:26:18Z citable
#> 3   oai:pangaea.de:doi:10.1594/PANGAEA.57539 2017-03-05T14:25:32Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.142421 2017-03-05T13:53:45Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.149998 2017-03-05T13:59:10Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.208129 2017-03-05T14:35:04Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.228741 2017-03-05T13:59:25Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.547798 2017-03-05T13:40:00Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.547942 2017-03-05T13:40:00Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.552177 2017-03-05T14:25:31Z citable
#> # ... with 19,227 more rows, and 4 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, setSpec.3 <chr>, setSpec.4 <chr>

pg_list_identifiers(set="geocode1", from=Sys.Date()-1, until=Sys.Date())

#> # A tibble: 351 × 8
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1  oai:pangaea.de:doi:10.1594/PANGAEA.314690 2017-03-07T11:47:26Z citable
#> 2  oai:pangaea.de:doi:10.1594/PANGAEA.611095 2017-03-07T11:01:44Z citable
#> 3  oai:pangaea.de:doi:10.1594/PANGAEA.716835 2017-03-07T10:52:54Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.785456 2017-03-07T11:02:51Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.807446 2017-03-07T11:06:14Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.846306 2017-03-07T10:35:30Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.867475 2017-03-06T13:50:35Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.871636 2017-03-07T03:10:43Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.871704 2017-03-06T03:15:28Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.871714 2017-03-06T03:15:30Z citable
#> # ... with 341 more rows, and 5 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, setSpec.3 <chr>, setSpec.4 <chr>, setSpec.5 <chr>

pg_list_identifiers(prefix="iso19139", from=Sys.Date()-1, until=Sys.Date())

#> # A tibble: 1,450 × 7
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1  oai:pangaea.de:doi:10.1594/PANGAEA.149999 2017-03-07T09:22:22Z citable
#> 2  oai:pangaea.de:doi:10.1594/PANGAEA.314690 2017-03-07T11:47:26Z citable
#> 3  oai:pangaea.de:doi:10.1594/PANGAEA.611095 2017-03-07T11:01:44Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.716835 2017-03-07T10:52:54Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.728847 2017-03-07T13:17:12Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.735539 2017-03-07T13:19:15Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.737438 2017-03-06T15:40:27Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.745833 2017-03-07T13:19:13Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.746016 2017-03-07T09:22:25Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.753644 2017-03-07T11:44:13Z citable
#> # ... with 1,440 more rows, and 4 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, setSpec.3 <chr>, setSpec.4 <chr>

pg_list_identifiers(prefix="dif",
  from = paste0(Sys.Date() - 2, "T00:00:00Z"),
  until = paste0(Sys.Date() - 1, "T18:00:00Z")
)

#> # A tibble: 19,237 × 7
#>                                   identifier            datestamp setSpec
#>                                        <chr>                <chr>   <chr>
#> 1   oai:pangaea.de:doi:10.1594/PANGAEA.52692 2017-03-05T14:44:38Z citable
#> 2   oai:pangaea.de:doi:10.1594/PANGAEA.53178 2017-03-05T14:26:18Z citable
#> 3   oai:pangaea.de:doi:10.1594/PANGAEA.57539 2017-03-05T14:25:32Z citable
#> 4  oai:pangaea.de:doi:10.1594/PANGAEA.142421 2017-03-05T13:53:45Z citable
#> 5  oai:pangaea.de:doi:10.1594/PANGAEA.149998 2017-03-05T13:59:10Z citable
#> 6  oai:pangaea.de:doi:10.1594/PANGAEA.208129 2017-03-05T14:35:04Z citable
#> 7  oai:pangaea.de:doi:10.1594/PANGAEA.228741 2017-03-05T13:59:25Z citable
#> 8  oai:pangaea.de:doi:10.1594/PANGAEA.547798 2017-03-05T13:40:00Z citable
#> 9  oai:pangaea.de:doi:10.1594/PANGAEA.547942 2017-03-05T13:40:00Z citable
#> 10 oai:pangaea.de:doi:10.1594/PANGAEA.552177 2017-03-05T14:25:31Z citable
#> # ... with 19,227 more rows, and 4 more variables: setSpec.1 <chr>,
#> #   setSpec.2 <chr>, setSpec.3 <chr>, setSpec.4 <chr>