Search the Pangaea database

Search the Pangaea database

pg_search(query, count = 10, offset = 0, topic = NULL, bbox = NULL,
  mindate = NULL, maxdate = NULL, ...)

Arguments

query

(character) Query terms. You can refine a search by prefixing the term(s) with a category, one of citation, reference, parameter, event, project, campaign, or basis. See examples.

count

(integer) Number of items to return. Default: 10. Maximum: 500. Use offset parameter to page through results - see examples

offset

(integer) Record number to start at. Default: 0

topic

(character) topic area: one of NULL (all areas), "Agriculture", "Atomosphere", "Biological Classification", "Biospshere", "Chemistry", "Cryosphere", "Ecology", "Fisheries", "Geophysics", "Human Dimensions", "Lakes & Rivers", "Land Surface", "Lithosphere", "Oceans", "Paleontology"

bbox

(numeric) A bounding box, of the form: minlon, minlat, maxlon, maxlat

mindate, maxdate

(character) Dates to search for, of the form "2014-10-28"

...

Curl options passed on to httr::GET()

Value

tibble/data.frame with the structure:

score - match score, higher is a better match

doi - the DOI for the data package

size - size number

size_measure - size measure, one of "data points" or "datasets"

citation - citation for the data package

supplement_to - citation for what the data package is a supplement to

Details

This is a thin wrapper around the GUI search interface on the page https://www.pangaea.de. Everything you can do there, you can do here.

See also

pg_search_es()

Examples

pg_search(query='water')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  26.31440 10.1594/PANGAEA.843740     7     datasets
#> 2  26.26142 10.1594/PANGAEA.762952   127  data points
#> 3  26.15477 10.1594/PANGAEA.811079    10     datasets
#> 4  26.11712 10.1594/PANGAEA.759113   496  data points
#> 5  26.10503 10.1594/PANGAEA.825428     5     datasets
#> 6  25.97657 10.1594/PANGAEA.738090     2     datasets
#> 7  25.85503 10.1594/PANGAEA.862525   372     datasets
#> 8  25.84507 10.1594/PANGAEA.738239     8     datasets
#> 9  25.73387 10.1594/PANGAEA.806957    15  data points
#> 10 25.68901 10.1594/PANGAEA.812094     2     datasets
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='water', count=2)

#> # A tibble: 2 × 6
#>      score                    doi  size size_measure
#> *    <dbl>                  <chr> <dbl>        <chr>
#> 1 26.31440 10.1594/PANGAEA.843740     7     datasets
#> 2 26.26142 10.1594/PANGAEA.762952   127  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='water', count=20)

#> # A tibble: 20 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  26.31440 10.1594/PANGAEA.843740     7     datasets
#> 2  26.26142 10.1594/PANGAEA.762952   127  data points
#> 3  26.15477 10.1594/PANGAEA.811079    10     datasets
#> 4  26.11712 10.1594/PANGAEA.759113   496  data points
#> 5  26.10503 10.1594/PANGAEA.825428     5     datasets
#> 6  25.97657 10.1594/PANGAEA.738090     2     datasets
#> 7  25.85503 10.1594/PANGAEA.862525   372     datasets
#> 8  25.84507 10.1594/PANGAEA.738239     8     datasets
#> 9  25.73387 10.1594/PANGAEA.806957    15  data points
#> 10 25.68901 10.1594/PANGAEA.812094     2     datasets
#> 11 25.42802 10.1594/PANGAEA.807926     2     datasets
#> 12 23.63116 10.1594/PANGAEA.773526     6     datasets
#> 13 23.29432 10.1594/PANGAEA.784449    11     datasets
#> 14 23.27849 10.1594/PANGAEA.735614  3600  data points
#> 15 23.01474 10.1594/PANGAEA.738279     4     datasets
#> 16 22.79517 10.1594/PANGAEA.860500  4114       kBytes
#> 17 22.62421 10.1594/PANGAEA.804382   997  data points
#> 18 22.45825 10.1594/PANGAEA.778629   730  data points
#> 19 22.26147 10.1594/PANGAEA.667860   878  data points
#> 20 21.80358 10.1594/PANGAEA.846281    12  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='water', mindate="2013-06-01", maxdate="2013-07-01")

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  25.85503 10.1594/PANGAEA.862525   372     datasets
#> 2  21.04407 10.1594/PANGAEA.831007    54     datasets
#> 3  20.63197 10.1594/PANGAEA.854921    62     datasets
#> 4  15.35937 10.1594/PANGAEA.859527    20     datasets
#> 5  15.30176 10.1594/PANGAEA.150009   100  data points
#> 6  14.45300 10.1594/PANGAEA.833840     3     datasets
#> 7  14.42722 10.1594/PANGAEA.845181     2     datasets
#> 8  14.29151 10.1594/PANGAEA.855825   475  data points
#> 9  14.07505 10.1594/PANGAEA.864251     3     datasets
#> 10 14.03037 10.1594/PANGAEA.845938     3     datasets
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='water', bbox=c(-124.2, 41.8, -116.8, 46.1))

#> # A tibble: 10 × 6
#>        score                    doi  size size_measure
#> *      <dbl>                  <chr> <dbl>        <chr>
#> 1  21.585087 10.1594/PANGAEA.812094     2     datasets
#> 2  11.017511 10.1594/PANGAEA.862525   372     datasets
#> 3  10.910356 10.1594/PANGAEA.736010     9     datasets
#> 4  10.352143 10.1594/PANGAEA.803591     2     datasets
#> 5  10.287091 10.1594/PANGAEA.707903     2     datasets
#> 6   8.576023 10.1594/PANGAEA.774780   448  data points
#> 7   8.239835 10.1594/PANGAEA.406622   433  data points
#> 8   8.238589 10.1594/PANGAEA.401390   500  data points
#> 9   8.231977 10.1594/PANGAEA.402042   392  data points
#> 10  8.230261 10.1594/PANGAEA.406619   480  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='reference:Archer')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  15.97091 10.1594/PANGAEA.832860     2     datasets
#> 2  15.97091 10.1594/PANGAEA.815226     2     datasets
#> 3  13.70649 10.1594/PANGAEA.107286   116       kBytes
#> 4  13.70649  10.1594/PANGAEA.56017 36229  data points
#> 5  11.74025 10.1594/PANGAEA.121082    16  data points
#> 6  11.74025 10.1594/PANGAEA.121162    80  data points
#> 7  11.74025 10.1594/PANGAEA.121239    20  data points
#> 8  11.74025 10.1594/PANGAEA.121158    32  data points
#> 9  11.74025 10.1594/PANGAEA.121222    32  data points
#> 10 11.74025 10.1594/PANGAEA.121144   135  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='parameter:"carbon dioxide"')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  40.27695 10.1594/PANGAEA.854845    88  data points
#> 2  38.73645 10.1594/PANGAEA.710905   219  data points
#> 3  38.73645 10.1594/PANGAEA.666157  1116  data points
#> 4  38.73645 10.1594/PANGAEA.666115  2332  data points
#> 5  38.33087 10.1594/PANGAEA.809527   162  data points
#> 6  37.47801 10.1594/PANGAEA.135711 35234  data points
#> 7  37.39484 10.1594/PANGAEA.800072  1886  data points
#> 8  37.39484 10.1594/PANGAEA.853833    35  data points
#> 9  36.38170 10.1594/PANGAEA.859395   123  data points
#> 10 36.38170 10.1594/PANGAEA.707419    20  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='event:M2-track')

#> # A tibble: 6 × 6
#>      score                    doi  size size_measure
#> *    <dbl>                  <chr> <dbl>        <chr>
#> 1 64.51608 10.1594/PANGAEA.666397  3362  data points
#> 2 64.51608 10.1594/PANGAEA.666115  2332  data points
#> 3 61.05290 10.1594/PANGAEA.666344   497  data points
#> 4 61.05290 10.1594/PANGAEA.665707   254  data points
#> 5 20.35097 10.1594/PANGAEA.680864 30483  data points
#> 6 20.35097 10.1594/PANGAEA.687643  6150  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='event:TT011_2-CTD31')

#> # A tibble: 3 × 6
#>        score                    doi  size size_measure
#> *      <dbl>                  <chr> <dbl>        <chr>
#> 1 24.4086250 10.1594/PANGAEA.121189    45  data points
#> 2 23.3173300 10.1594/PANGAEA.120139  1476  data points
#> 3  0.7133222 10.1594/PANGAEA.119802   991  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <lgl>

pg_search(query='project:Joint Global Ocean Flux Study')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  211.9782 10.1594/PANGAEA.760907   607       MBytes
#> 2  141.6526 10.1594/PANGAEA.734159     4     datasets
#> 3  125.2556 10.1594/PANGAEA.807661     5     datasets
#> 4  123.2449 10.1594/PANGAEA.760086     5     datasets
#> 5  121.7838 10.1594/PANGAEA.807672     4     datasets
#> 6  120.8107 10.1594/PANGAEA.787478    36     datasets
#> 7  120.5973 10.1594/PANGAEA.742668     6     datasets
#> 8  116.9698 10.1594/PANGAEA.729929    10     datasets
#> 9  113.1684 10.1594/PANGAEA.728676   153     datasets
#> 10 112.0901 10.1594/PANGAEA.807617   864  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='campaign:M2')

#> # A tibble: 10 × 6
#>        score                    doi  size size_measure
#> *      <dbl>                  <chr> <dbl>        <chr>
#> 1  13.088648 10.1594/PANGAEA.666344   497  data points
#> 2  13.088648 10.1594/PANGAEA.548400     6     datasets
#> 3  13.088648 10.1594/PANGAEA.665707   254  data points
#> 4  12.709326 10.1594/PANGAEA.666397  3362  data points
#> 5  12.709326 10.1594/PANGAEA.666115  2332  data points
#> 6  12.709326 10.1594/PANGAEA.663168   335     datasets
#> 7  12.709326 10.1594/PANGAEA.665700   648  data points
#> 8   8.809006 10.1594/PANGAEA.603892    22     datasets
#> 9   8.809006 10.1594/PANGAEA.666396    42     datasets
#> 10  4.362883 10.1594/PANGAEA.680864 30483  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='basis:Meteor')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  13.62580  10.1594/PANGAEA.55046 24570  data points
#> 2  11.99645 10.1594/PANGAEA.696097     7     datasets
#> 3  11.95508 10.1594/PANGAEA.743064     2     datasets
#> 4  11.38526 10.1594/PANGAEA.763395     8  data points
#> 5  11.38526 10.1594/PANGAEA.761653 14852  data points
#> 6  11.38526 10.1594/PANGAEA.779104     4     datasets
#> 7  11.38526 10.1594/PANGAEA.611330    40     datasets
#> 8  11.38526 10.1594/PANGAEA.738495    11     datasets
#> 9  11.38526 10.1594/PANGAEA.726626     2     datasets
#> 10 11.38526 10.1594/PANGAEA.604012     6     datasets
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

# paging with count and offset
# max is 500 records per request - if you need > 500, use offset and count
res1 <- pg_search(query = "florisphaera", count = 500, offset = 0)
res2 <- pg_search(query = "florisphaera", count = 500, offset = 500)
res3 <- pg_search(query = "florisphaera", count = 500, offset = 1000)
do.call("rbind.data.frame", list(res1, res2, res3))

#> # A tibble: 1,235 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  79.11172 10.1594/PANGAEA.805696     2     datasets
#> 2  54.37842 10.1594/PANGAEA.757616   150  data points
#> 3  53.64798 10.1594/PANGAEA.122790   852  data points
#> 4  47.01232 10.1594/PANGAEA.836238  1946  data points
#> 5  47.01232 10.1594/PANGAEA.867495   416  data points
#> 6  44.68165 10.1594/PANGAEA.816226  3744  data points
#> 7  44.67970 10.1594/PANGAEA.864106     4     datasets
#> 8  43.96087 10.1594/PANGAEA.846603   162  data points
#> 9  43.91185 10.1594/PANGAEA.844527     7     datasets
#> 10 41.74735 10.1594/PANGAEA.746273  1728  data points
#> # ... with 1,225 more rows, and 2 more variables: citation <chr>,
#> #   supplement_to <chr>

# get attributes: maxScore, totalCount, and offset
res <- pg_search(query='water', bbox=c(-124.2, 41.8, -116.8, 46.1))
attributes(res)

#> $names
#> [1] "score"         "doi"           "size"          "size_measure" 
#> [5] "citation"      "supplement_to"
#> 
#> $row.names
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $class
#> [1] "tbl_df"     "tbl"        "data.frame"
#> 
#> $maxScore
#> [1] 21.58509
#> 
#> $offset
#> [1] 0
#> 
#> $totalCount
#> [1] 610
#> 

attr(res, "maxScore")

#> [1] 21.58509

attr(res, "totalCount")

#> [1] 610

attr(res, "offset")

#> [1] 0

# curl options
library(httr)
pg_search(query='citation:Archer', config = verbose())

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  28.34635 10.1594/PANGAEA.758219   285  data points
#> 2  21.22835 10.1594/PANGAEA.730420    33     datasets
#> 3  21.22835 10.1594/PANGAEA.736010     9     datasets
#> 4  21.22835 10.1594/PANGAEA.730527    21     datasets
#> 5  21.22835 10.1594/PANGAEA.726364     4     datasets
#> 6  21.22835 10.1594/PANGAEA.715051     5     datasets
#> 7  21.22835 10.1594/PANGAEA.721741   978  data points
#> 8  14.20872 10.1594/PANGAEA.121082    16  data points
#> 9  14.20872 10.1594/PANGAEA.121162    80  data points
#> 10 14.20872 10.1594/PANGAEA.121239    20  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>