Search the Pangaea database

Search the Pangaea database

pg_search(query, count = 10, offset = 0, topic = NULL, bbox = NULL,
  mindate = NULL, maxdate = NULL, ...)

Arguments

query

(character) Query terms. You can refine a search by prefixing the term(s) with a category, one of citation, reference, parameter, event, project, campaign, or basis. See examples.

count

(integer) Number of items to return. Default: 10. Maximum: 500. Use offset parameter to page through results - see examples

offset

(integer) Record number to start at. Default: 0

topic

(character) topic area: one of NULL (all areas), "Agriculture", "Atomosphere", "Biological Classification", "Biospshere", "Chemistry", "Cryosphere", "Ecology", "Fisheries", "Geophysics", "Human Dimensions", "Lakes & Rivers", "Land Surface", "Lithosphere", "Oceans", "Paleontology"

bbox

(numeric) A bounding box, of the form: minlon, minlat, maxlon, maxlat

mindate, maxdate

(character) Dates to search for, of the form "2014-10-28"

...

Curl options passed on to httr::GET()

Value

tibble/data.frame with the structure:

score - match score, higher is a better match

doi - the DOI for the data package

size - size number

size_measure - size measure, one of "data points" or "datasets"

citation - citation for the data package

supplement_to - citation for what the data package is a supplement to

Details

This is a thin wrapper around the GUI search interface on the page https://www.pangaea.de. Everything you can do there, you can do here.

See also

pg_search_es()

Examples

pg_search(query='water')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  26.60823 10.1594/PANGAEA.843740     7     datasets
#> 2  26.44750 10.1594/PANGAEA.811079    10     datasets
#> 3  26.44294 10.1594/PANGAEA.762952   127  data points
#> 4  26.40995 10.1594/PANGAEA.759113   496  data points
#> 5  26.28622 10.1594/PANGAEA.825428     5     datasets
#> 6  26.16394 10.1594/PANGAEA.738090     2     datasets
#> 7  26.04250 10.1594/PANGAEA.862525   372     datasets
#> 8  26.03212 10.1594/PANGAEA.738239     8     datasets
#> 9  26.01679 10.1594/PANGAEA.806957    15  data points
#> 10 25.97286 10.1594/PANGAEA.812094     2     datasets
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='water', count=2)

#> # A tibble: 2 × 6
#>      score                    doi  size size_measure
#> *    <dbl>                  <chr> <dbl>        <chr>
#> 1 26.60823 10.1594/PANGAEA.843740     7     datasets
#> 2 26.44750 10.1594/PANGAEA.811079    10     datasets
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='water', count=20)

#> # A tibble: 20 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  26.60823 10.1594/PANGAEA.843740     7     datasets
#> 2  26.44750 10.1594/PANGAEA.811079    10     datasets
#> 3  26.44294 10.1594/PANGAEA.762952   127  data points
#> 4  26.40995 10.1594/PANGAEA.759113   496  data points
#> 5  26.28622 10.1594/PANGAEA.825428     5     datasets
#> 6  26.16394 10.1594/PANGAEA.738090     2     datasets
#> 7  26.04250 10.1594/PANGAEA.862525   372     datasets
#> 8  26.03212 10.1594/PANGAEA.738239     8     datasets
#> 9  26.01679 10.1594/PANGAEA.806957    15  data points
#> 10 25.97286 10.1594/PANGAEA.812094     2     datasets
#> 11 25.60824 10.1594/PANGAEA.807926     2     datasets
#> 12 23.81278 10.1594/PANGAEA.773526     6     datasets
#> 13 23.57081 10.1594/PANGAEA.784449    11     datasets
#> 14 23.45975 10.1594/PANGAEA.735614  3600  data points
#> 15 23.29199 10.1594/PANGAEA.738279     4     datasets
#> 16 22.98897 10.1594/PANGAEA.860500  4114       kBytes
#> 17 22.81498 10.1594/PANGAEA.804382   997  data points
#> 18 22.57957 10.1594/PANGAEA.778629   730  data points
#> 19 22.38134 10.1594/PANGAEA.667860   878  data points
#> 20 21.98198 10.1594/PANGAEA.761833  7920  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='water', mindate="2013-06-01", maxdate="2013-07-01")

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  26.04250 10.1594/PANGAEA.862525   372     datasets
#> 2  21.21439 10.1594/PANGAEA.831007    54     datasets
#> 3  20.80053 10.1594/PANGAEA.854921    62     datasets
#> 4  15.45649 10.1594/PANGAEA.859527    20     datasets
#> 5  14.61787 10.1594/PANGAEA.845181     2     datasets
#> 6  14.61712 10.1594/PANGAEA.833840     3     datasets
#> 7  14.61632 10.1594/PANGAEA.150009   100  data points
#> 8  14.48137 10.1594/PANGAEA.855825   475  data points
#> 9  14.20332 10.1594/PANGAEA.864251     3     datasets
#> 10 14.05052 10.1594/PANGAEA.845938     3     datasets
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='water', bbox=c(-124.2, 41.8, -116.8, 46.1))

#> # A tibble: 10 × 6
#>        score                    doi  size size_measure
#> *      <dbl>                  <chr> <dbl>        <chr>
#> 1  21.823591 10.1594/PANGAEA.812094     2     datasets
#> 2  11.097401 10.1594/PANGAEA.862525   372     datasets
#> 3  10.921976 10.1594/PANGAEA.736010     9     datasets
#> 4  10.422461 10.1594/PANGAEA.707903     2     datasets
#> 5  10.375950 10.1594/PANGAEA.803591     2     datasets
#> 6   8.574010 10.1594/PANGAEA.774780   448  data points
#> 7   8.326280 10.1594/PANGAEA.406622   433  data points
#> 8   8.325020 10.1594/PANGAEA.401390   500  data points
#> 9   8.316605 10.1594/PANGAEA.406619   480  data points
#> 10  8.316592 10.1594/PANGAEA.406110   598  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='reference:Archer')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  16.00030 10.1594/PANGAEA.832860     2     datasets
#> 2  16.00030 10.1594/PANGAEA.815226     2     datasets
#> 3  13.73814 10.1594/PANGAEA.107286   116       kBytes
#> 4  13.73814  10.1594/PANGAEA.56017 36229  data points
#> 5  11.78526 10.1594/PANGAEA.121082    16  data points
#> 6  11.78526 10.1594/PANGAEA.121162    80  data points
#> 7  11.78526 10.1594/PANGAEA.121239    20  data points
#> 8  11.78526 10.1594/PANGAEA.121158    32  data points
#> 9  11.78526 10.1594/PANGAEA.121222    32  data points
#> 10 11.78526 10.1594/PANGAEA.121144   135  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='parameter:"carbon dioxide"')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  40.40111 10.1594/PANGAEA.854845    88  data points
#> 2  38.86488 10.1594/PANGAEA.710905   219  data points
#> 3  38.86488 10.1594/PANGAEA.666157  1116  data points
#> 4  38.86488 10.1594/PANGAEA.666115  2332  data points
#> 5  38.43412 10.1594/PANGAEA.809527   162  data points
#> 6  37.57910 10.1594/PANGAEA.135711 35234  data points
#> 7  37.53632 10.1594/PANGAEA.800072  1886  data points
#> 8  37.53632 10.1594/PANGAEA.853833    35  data points
#> 9  36.52519 10.1594/PANGAEA.859395   123  data points
#> 10 36.52519 10.1594/PANGAEA.707419    20  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='event:M2-track')

#> # A tibble: 6 × 6
#>      score                    doi  size size_measure
#> *    <dbl>                  <chr> <dbl>        <chr>
#> 1 64.37206 10.1594/PANGAEA.666397  3362  data points
#> 2 64.37206 10.1594/PANGAEA.666115  2332  data points
#> 3 61.11130 10.1594/PANGAEA.666344   497  data points
#> 4 61.11130 10.1594/PANGAEA.665707   254  data points
#> 5 20.37043 10.1594/PANGAEA.680864 30483  data points
#> 6 20.37043 10.1594/PANGAEA.687643  6150  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='event:TT011_2-CTD31')

#> # A tibble: 3 × 6
#>        score                    doi  size size_measure
#> *      <dbl>                  <chr> <dbl>        <chr>
#> 1 24.3646200 10.1594/PANGAEA.121189    45  data points
#> 2 23.3394970 10.1594/PANGAEA.120139  1476  data points
#> 3  0.7134453 10.1594/PANGAEA.119802   991  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <lgl>

pg_search(query='project:Joint Global Ocean Flux Study')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  213.8270 10.1594/PANGAEA.760907   607       MBytes
#> 2  142.6300 10.1594/PANGAEA.734159     4     datasets
#> 3  126.2488 10.1594/PANGAEA.807661     5     datasets
#> 4  124.0989 10.1594/PANGAEA.760086     5     datasets
#> 5  122.7366 10.1594/PANGAEA.807672     4     datasets
#> 6  121.6255 10.1594/PANGAEA.787478    36     datasets
#> 7  121.4493 10.1594/PANGAEA.742668     6     datasets
#> 8  118.2036 10.1594/PANGAEA.729929    10     datasets
#> 9  114.0571 10.1594/PANGAEA.728676   153     datasets
#> 10 112.9133 10.1594/PANGAEA.807617   864  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='campaign:M2')

#> # A tibble: 10 × 6
#>        score                    doi  size size_measure
#> *      <dbl>                  <chr> <dbl>        <chr>
#> 1  13.091860 10.1594/PANGAEA.666344   497  data points
#> 2  13.091860 10.1594/PANGAEA.548400     6     datasets
#> 3  13.091860 10.1594/PANGAEA.665707   254  data points
#> 4  12.690264 10.1594/PANGAEA.666397  3362  data points
#> 5  12.690264 10.1594/PANGAEA.666115  2332  data points
#> 6  12.690264 10.1594/PANGAEA.663168   335     datasets
#> 7  12.690264 10.1594/PANGAEA.665700   648  data points
#> 8   8.794107 10.1594/PANGAEA.603892    22     datasets
#> 9   8.794107 10.1594/PANGAEA.666396    42     datasets
#> 10  4.363953 10.1594/PANGAEA.680864 30483  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

pg_search(query='basis:Meteor')

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  13.63702  10.1594/PANGAEA.55046 24570  data points
#> 2  12.00613 10.1594/PANGAEA.696097     7     datasets
#> 3  11.91643 10.1594/PANGAEA.743064     2     datasets
#> 4  11.39455 10.1594/PANGAEA.763395     8  data points
#> 5  11.39455 10.1594/PANGAEA.761653 14852  data points
#> 6  11.39455 10.1594/PANGAEA.779104     4     datasets
#> 7  11.39455 10.1594/PANGAEA.611330    40     datasets
#> 8  11.39455 10.1594/PANGAEA.738495    11     datasets
#> 9  11.39455 10.1594/PANGAEA.726626     2     datasets
#> 10 11.39455 10.1594/PANGAEA.604012     6     datasets
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>

# paging with count and offset
# max is 500 records per request - if you need > 500, use offset and count
res1 <- pg_search(query = "florisphaera", count = 500, offset = 0)
res2 <- pg_search(query = "florisphaera", count = 500, offset = 500)
res3 <- pg_search(query = "florisphaera", count = 500, offset = 1000)
do.call("rbind.data.frame", list(res1, res2, res3))

#> # A tibble: 1,235 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  79.63031 10.1594/PANGAEA.805696     2     datasets
#> 2  54.67973 10.1594/PANGAEA.757616   150  data points
#> 3  53.94842 10.1594/PANGAEA.122790   852  data points
#> 4  47.37415 10.1594/PANGAEA.836238  1946  data points
#> 5  47.37415 10.1594/PANGAEA.867495   416  data points
#> 6  44.60737 10.1594/PANGAEA.816226  3744  data points
#> 7  44.60462 10.1594/PANGAEA.864106     4     datasets
#> 8  43.88036 10.1594/PANGAEA.846603   162  data points
#> 9  43.83250 10.1594/PANGAEA.844527     7     datasets
#> 10 41.96066 10.1594/PANGAEA.691005   510  data points
#> # ... with 1,225 more rows, and 2 more variables: citation <chr>,
#> #   supplement_to <chr>

# get attributes: maxScore, totalCount, and offset
res <- pg_search(query='water', bbox=c(-124.2, 41.8, -116.8, 46.1))
attributes(res)

#> $names
#> [1] "score"         "doi"           "size"          "size_measure" 
#> [5] "citation"      "supplement_to"
#> 
#> $row.names
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $class
#> [1] "tbl_df"     "tbl"        "data.frame"
#> 
#> $maxScore
#> [1] 21.82359
#> 
#> $offset
#> [1] 0
#> 
#> $totalCount
#> [1] 610
#> 

attr(res, "maxScore")

#> [1] 21.82359

attr(res, "totalCount")

#> [1] 610

attr(res, "offset")

#> [1] 0

# curl options
library(httr)
pg_search(query='citation:Archer', config = verbose())

#> # A tibble: 10 × 6
#>       score                    doi  size size_measure
#> *     <dbl>                  <chr> <dbl>        <chr>
#> 1  28.60838 10.1594/PANGAEA.758219   285  data points
#> 2  21.45116 10.1594/PANGAEA.730420    33     datasets
#> 3  21.45116 10.1594/PANGAEA.736010     9     datasets
#> 4  21.45116 10.1594/PANGAEA.730527    21     datasets
#> 5  21.45116 10.1594/PANGAEA.721741   978  data points
#> 6  21.45116 10.1594/PANGAEA.715051     5     datasets
#> 7  21.45116 10.1594/PANGAEA.726364     4     datasets
#> 8  14.30743 10.1594/PANGAEA.121082    16  data points
#> 9  14.30743 10.1594/PANGAEA.121162    80  data points
#> 10 14.30743 10.1594/PANGAEA.121239    20  data points
#> # ... with 2 more variables: citation <chr>, supplement_to <chr>