Search the Pangaea database with Elasticsearch

Search the Pangaea database with Elasticsearch

pg_search_es(query = NULL, size = 10, from = NULL, source = NULL,
  df = NULL, analyzer = NULL, default_operator = NULL, explain = NULL,
  sort = NULL, track_scores = NULL, timeout = NULL,
  terminate_after = NULL, search_type = NULL,
  lowercase_expanded_terms = NULL, analyze_wildcard = NULL,
  version = FALSE, ...)

Arguments

query

(character) Query terms..

size

(character) The number of hits to return. Pass in as a character string to avoid problems with large number conversion to scientific notation. Default: 10. The default maximum is 10,000 - however, you can change this default maximum by changing the index.max_result_window index level parameter.

from

(character) The starting from index of the hits to return. Pass in as a character string to avoid problems with large number conversion to scientific notation. Default: 0

source

(character) character vector of fields to return

df

(character) The default field to use when no field prefix is defined within the query.

analyzer

(character) The analyzer name to be used when analyzing the query string.

default_operator

(character) The default operator to be used, can be AND or OR. Default: OR

explain

(logical) For each hit, contain an explanation of how scoring of the hits was computed. Default: FALSE

sort

(character) Sorting to perform. Can either be in the form of fieldName, or fieldName:asc/fieldName:desc. The fieldName can either be an actual field within the document, or the special _score name to indicate sorting based on scores. There can be several sort parameters (order is important).

track_scores

(logical) When sorting, set to TRUE in order to still track scores and return them as part of each hit.

timeout

(numeric) A search timeout, bounding the search request to be executed within the specified time value and bail with the hits accumulated up to that point when expired. Default: no timeout.

terminate_after

(numeric) The maximum number of documents to collect for each shard, upon reaching which the query execution will terminate early. If set, the response will have a boolean field terminated_early to indicate whether the query execution has actually terminated_early. Default: no terminate_after

search_type

(character) The type of the search operation to perform. Can be query_then_fetch (default) or dfs_query_then_fetch. Types scan and count are deprecated. See http://bit.ly/19Am9xP for more details on the different types of search that can be performed.

lowercase_expanded_terms

(logical) Should terms be automatically lowercased or not. Default: TRUE.

analyze_wildcard

(logical) Should wildcard and prefix queries be analyzed or not. Default: FALSE

version

(logical) Print the document version with each document.

...

Curl options passed on to httr::GET()

Value

tibble/data.frame, empty if no results

Details

An interface to Pangaea's Elasticsearch query interface. You can also just use https://github.com/ropensci/elastic package to interact with it. The base URL is https://ws.pangaea.de/es/pangaea/panmd/_search

See also

pg_search()

Examples

(res <- pg_search_es())

#> total hits: 361563

#> # A tibble: 10 × 29
#>      `_index` `_type`  `_id` `_score` `_source.internal-datestamp`
#> *       <chr>   <chr>  <chr>    <dbl>                        <chr>
#> 1  pangaea_v2   panmd 578301        1     2016-10-11T09:54:31.000Z
#> 2  pangaea_v2   panmd 575333        1     2016-10-11T09:55:35.000Z
#> 3  pangaea_v2   panmd 606983        1     2016-10-11T10:24:04.000Z
#> 4  pangaea_v2   panmd 578422        1     2016-10-11T09:54:24.000Z
#> 5  pangaea_v2   panmd 575742        1     2016-10-11T09:55:29.000Z
#> 6  pangaea_v2   panmd 605954        1     2016-10-11T10:22:39.000Z
#> 7  pangaea_v2   panmd 558380        1     2016-10-11T08:25:10.000Z
#> 8  pangaea_v2   panmd 575470        1     2016-10-11T08:21:41.000Z
#> 9  pangaea_v2   panmd 575601        1     2016-10-11T09:55:31.000Z
#> 10 pangaea_v2   panmd 604994        1     2016-10-11T10:26:47.000Z
#> # ... with 24 more variables: `_source.minElevation` <dbl>,
#> #   `_source.sf-authortitle` <chr>, `_source.xml-sitemap` <chr>,
#> #   `_source.techKeyword` <list>, `_source.geocodes` <list>,
#> #   `_source.sp-loginOption` <int>, `_source.westBoundLongitude` <dbl>,
#> #   `_source.eastBoundLongitude` <dbl>, `_source.URI` <chr>,
#> #   `_source.northBoundLatitude` <dbl>, `_source.sp-dataStatus` <int>,
#> #   `_source.xml-thumb` <chr>, `_source.sp-hidden` <int>, `_source.xml` <chr>,
#> #   `_source.internal-source` <chr>, `_source.southBoundLatitude` <dbl>,
#> #   `_source.idDataSet` <int>, `_source.boost` <dbl>,
#> #   `_source.maxElevation` <dbl>, `_source.parentURI` <chr>,
#> #   `_source.parentIdDataSet` <int>, `_source.oaiSet` <chr>,
#> #   `_source.meanPosition.lat` <dbl>, `_source.meanPosition.lon` <dbl>

attributes(res)

#> $names
#>  [1] "_index"                     "_type"                     
#>  [3] "_id"                        "_score"                    
#>  [5] "_source.internal-datestamp" "_source.minElevation"      
#>  [7] "_source.sf-authortitle"     "_source.xml-sitemap"       
#>  [9] "_source.techKeyword"        "_source.geocodes"          
#> [11] "_source.sp-loginOption"     "_source.westBoundLongitude"
#> [13] "_source.eastBoundLongitude" "_source.URI"               
#> [15] "_source.northBoundLatitude" "_source.sp-dataStatus"     
#> [17] "_source.xml-thumb"          "_source.sp-hidden"         
#> [19] "_source.xml"                "_source.internal-source"   
#> [21] "_source.southBoundLatitude" "_source.idDataSet"         
#> [23] "_source.boost"              "_source.maxElevation"      
#> [25] "_source.parentURI"          "_source.parentIdDataSet"   
#> [27] "_source.oaiSet"             "_source.meanPosition.lat"  
#> [29] "_source.meanPosition.lon"  
#> 
#> $row.names
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> 
#> $class
#> [1] "tbl_df"     "tbl"        "data.frame"
#> 
#> $total
#> [1] 361563
#> 
#> $max_score
#> [1] 1
#> 

attr(res, "total")

#> [1] 361563

attr(res, "max_score")

#> [1] 1

pg_search_es(query = 'water', source = c('parentURI', 'minElevation'))

#> total hits: 141207

#> # A tibble: 10 × 6
#>      `_index` `_type`  `_id` `_score`        `_source.parentURI`
#> *       <chr>   <chr>  <chr>    <dbl>                      <chr>
#> 1  pangaea_v2   panmd 810858 1.973378 doi:10.1594/PANGAEA.819717
#> 2  pangaea_v2   panmd 810857 1.968188 doi:10.1594/PANGAEA.819717
#> 3  pangaea_v2   panmd  55746 1.909301                       <NA>
#> 4  pangaea_v2   panmd  55747 1.892445                       <NA>
#> 5  pangaea_v2   panmd 873072 1.892088 doi:10.1594/PANGAEA.873075
#> 6  pangaea_v2   panmd 831466 1.883156 doi:10.1594/PANGAEA.831469
#> 7  pangaea_v2   panmd 854443 1.870381 doi:10.1594/PANGAEA.854468
#> 8  pangaea_v2   panmd 854445 1.870381 doi:10.1594/PANGAEA.854468
#> 9  pangaea_v2   panmd 873052 1.870147 doi:10.1594/PANGAEA.873054
#> 10 pangaea_v2   panmd 819717 1.869914                       <NA>
#> # ... with 1 more variables: `_source.minElevation` <dbl>

pg_search_es(query = 'water', size = 3)

#> total hits: 141207

#> # A tibble: 3 × 32
#>     `_index` `_type`  `_id` `_score` `_source.internal-datestamp`
#> *      <chr>   <chr>  <chr>    <dbl>                        <chr>
#> 1 pangaea_v2   panmd 810858 1.973378     2017-02-04T02:15:57.000Z
#> 2 pangaea_v2   panmd 810857 1.968188     2017-02-04T02:12:52.000Z
#> 3 pangaea_v2   panmd  55746 1.909301     2017-01-19T09:35:35.000Z
#> # ... with 27 more variables: `_source.parentURI` <chr>,
#> #   `_source.minElevation` <dbl>, `_source.sf-authortitle` <chr>,
#> #   `_source.parentIdDataSet` <int>, `_source.techKeyword` <list>,
#> #   `_source.geocodes` <list>, `_source.sp-loginOption` <int>,
#> #   `_source.eastBoundLongitude` <dbl>, `_source.URI` <chr>,
#> #   `_source.minDateTime` <chr>, `_source.xml-thumb` <chr>,
#> #   `_source.xml` <chr>, `_source.maxDateTime` <chr>,
#> #   `_source.xml-sitemap` <chr>, `_source.westBoundLongitude` <dbl>,
#> #   `_source.northBoundLatitude` <dbl>, `_source.sp-dataStatus` <int>,
#> #   `_source.sp-hidden` <int>, `_source.internal-source` <chr>,
#> #   `_source.southBoundLatitude` <dbl>, `_source.idDataSet` <int>,
#> #   `_source.boost` <dbl>, `_source.maxElevation` <dbl>,
#> #   `_source.elevationGeocode` <chr>, `_source.oaiSet` <chr>,
#> #   `_source.meanPosition.lat` <dbl>, `_source.meanPosition.lon` <dbl>

pg_search_es(query = 'water', size = 3, from = 10)

#> total hits: 141207

#> # A tibble: 3 × 32
#>     `_index` `_type`  `_id` `_score` `_source.internal-datestamp`
#> *      <chr>   <chr>  <chr>    <dbl>                        <chr>
#> 1 pangaea_v2   panmd 779032 1.869821     2017-02-24T08:49:38.000Z
#> 2 pangaea_v2   panmd 854442 1.865297     2017-01-30T15:42:07.000Z
#> 3 pangaea_v2   panmd 854444 1.865297     2017-01-30T15:42:08.000Z
#> # ... with 27 more variables: `_source.parentURI` <chr>,
#> #   `_source.minElevation` <dbl>, `_source.sf-authortitle` <chr>,
#> #   `_source.parentIdDataSet` <int>, `_source.techKeyword` <list>,
#> #   `_source.geocodes` <list>, `_source.sp-loginOption` <int>,
#> #   `_source.eastBoundLongitude` <dbl>, `_source.URI` <chr>,
#> #   `_source.minDateTime` <chr>, `_source.xml-thumb` <chr>,
#> #   `_source.xml` <chr>, `_source.maxDateTime` <chr>,
#> #   `_source.xml-sitemap` <chr>, `_source.westBoundLongitude` <dbl>,
#> #   `_source.northBoundLatitude` <dbl>, `_source.sp-dataStatus` <int>,
#> #   `_source.sp-hidden` <int>, `_source.internal-source` <chr>,
#> #   `_source.southBoundLatitude` <dbl>, `_source.idDataSet` <int>,
#> #   `_source.boost` <dbl>, `_source.oaiSet` <chr>,
#> #   `_source.maxElevation` <dbl>, `_source.elevationGeocode` <chr>,
#> #   `_source.meanPosition.lat` <dbl>, `_source.meanPosition.lon` <dbl>

pg_search_es(query = 'water sky', default_operator = "OR")

#> total hits: 141899

#> # A tibble: 10 × 33
#>      `_index` `_type`  `_id` `_score` `_source.internal-datestamp`
#> *       <chr>   <chr>  <chr>    <dbl>                        <chr>
#> 1  pangaea_v2   panmd 588218 13.58088     2016-10-08T17:30:18.000Z
#> 2  pangaea_v2   panmd 810739 13.54150     2016-12-10T05:21:27.000Z
#> 3  pangaea_v2   panmd 843208 13.54150     2016-06-25T09:44:20.000Z
#> 4  pangaea_v2   panmd 843210 13.32852     2016-06-25T09:44:20.000Z
#> 5  pangaea_v2   panmd 843211 13.32852     2016-06-25T09:44:21.000Z
#> 6  pangaea_v2   panmd 753832 13.21544     2016-06-25T11:51:37.000Z
#> 7  pangaea_v2   panmd 845444 13.03556     2016-06-25T15:07:20.000Z
#> 8  pangaea_v2   panmd 759690 12.31106     2016-06-25T14:49:54.000Z
#> 9  pangaea_v2   panmd 829437 10.77627     2016-06-24T19:41:41.000Z
#> 10 pangaea_v2   panmd 820115 10.77627     2016-06-25T01:52:35.000Z
#> # ... with 28 more variables: `_source.maxDateTime` <chr>,
#> #   `_source.minElevation` <dbl>, `_source.sf-authortitle` <chr>,
#> #   `_source.xml-sitemap` <chr>, `_source.techKeyword` <list>,
#> #   `_source.geocodes` <list>, `_source.sp-loginOption` <int>,
#> #   `_source.westBoundLongitude` <dbl>, `_source.eastBoundLongitude` <dbl>,
#> #   `_source.URI` <chr>, `_source.northBoundLatitude` <dbl>,
#> #   `_source.sp-dataStatus` <int>, `_source.minDateTime` <chr>,
#> #   `_source.xml-thumb` <chr>, `_source.sp-hidden` <int>, `_source.xml` <chr>,
#> #   `_source.internal-source` <chr>, `_source.southBoundLatitude` <dbl>,
#> #   `_source.idDataSet` <int>, `_source.boost` <dbl>,
#> #   `_source.maxElevation` <dbl>, `_source.elevationGeocode` <chr>,
#> #   `_source.parentURI` <chr>, `_source.parentIdDataSet` <int>,
#> #   `_source.oaiSet` <list>, `_source.supplementToDOI` <chr>,
#> #   `_source.meanPosition.lat` <dbl>, `_source.meanPosition.lon` <dbl>

pg_search_es(query = 'water sky', default_operator = "AND")

#> total hits: 43

#> # A tibble: 10 × 33
#>      `_index` `_type`  `_id`  `_score` `_source.internal-datestamp`
#> *       <chr>   <chr>  <chr>     <dbl>                        <chr>
#> 1  pangaea_v2   panmd 588218 13.580884     2016-10-08T17:30:18.000Z
#> 2  pangaea_v2   panmd 845444 13.035557     2016-06-25T15:07:20.000Z
#> 3  pangaea_v2   panmd 759690 12.311061     2016-06-25T14:49:54.000Z
#> 4  pangaea_v2   panmd 872596 10.686980     2017-02-22T19:20:56.000Z
#> 5  pangaea_v2   panmd 872598 10.686980     2017-02-22T13:26:55.000Z
#> 6  pangaea_v2   panmd 872044  9.531056     2017-02-21T13:39:06.000Z
#> 7  pangaea_v2   panmd 871490  9.106581     2017-02-28T09:43:01.000Z
#> 8  pangaea_v2   panmd 824787  9.085766     2016-06-25T00:06:55.000Z
#> 9  pangaea_v2   panmd 823527  9.085766     2016-06-24T22:21:18.000Z
#> 10 pangaea_v2   panmd 824826  9.085766     2016-06-24T23:02:19.000Z
#> # ... with 28 more variables: `_source.maxDateTime` <chr>,
#> #   `_source.minElevation` <dbl>, `_source.sf-authortitle` <chr>,
#> #   `_source.xml-sitemap` <chr>, `_source.techKeyword` <list>,
#> #   `_source.geocodes` <list>, `_source.sp-loginOption` <int>,
#> #   `_source.westBoundLongitude` <dbl>, `_source.eastBoundLongitude` <dbl>,
#> #   `_source.URI` <chr>, `_source.northBoundLatitude` <dbl>,
#> #   `_source.sp-dataStatus` <int>, `_source.minDateTime` <chr>,
#> #   `_source.xml-thumb` <chr>, `_source.sp-hidden` <int>, `_source.xml` <chr>,
#> #   `_source.internal-source` <chr>, `_source.southBoundLatitude` <dbl>,
#> #   `_source.idDataSet` <int>, `_source.boost` <dbl>,
#> #   `_source.maxElevation` <dbl>, `_source.elevationGeocode` <chr>,
#> #   `_source.supplementToDOI` <chr>, `_source.oaiSet` <list>,
#> #   `_source.parentURI` <chr>, `_source.parentIdDataSet` <int>,
#> #   `_source.meanPosition.lat` <dbl>, `_source.meanPosition.lon` <dbl>

pg_search_es(query = 'water', sort = "minElevation")

#> total hits: 141207

#> # A tibble: 10 × 30
#>      `_index` `_type`  `_id` `_score`      sort `_source.internal-datestamp`
#> *       <chr>   <chr>  <chr>    <lgl>    <list>                        <chr>
#> 1  pangaea_v2   panmd 775536       NA <dbl [1]>     2017-02-04T02:09:47.000Z
#> 2  pangaea_v2   panmd 754280       NA <dbl [1]>     2017-02-04T06:37:43.000Z
#> 3  pangaea_v2   panmd 736039       NA <dbl [1]>     2017-02-03T22:57:04.000Z
#> 4  pangaea_v2   panmd 736426       NA <dbl [1]>     2017-02-03T23:41:49.000Z
#> 5  pangaea_v2   panmd 736435       NA <dbl [1]>     2017-02-03T23:43:09.000Z
#> 6  pangaea_v2   panmd 736610       NA <dbl [1]>     2017-02-04T01:39:12.000Z
#> 7  pangaea_v2   panmd 736608       NA <dbl [1]>     2017-02-04T01:37:53.000Z
#> 8  pangaea_v2   panmd 736607       NA <dbl [1]>     2017-02-04T01:37:52.000Z
#> 9  pangaea_v2   panmd 736669       NA <dbl [1]>     2017-02-03T21:29:19.000Z
#> 10 pangaea_v2   panmd 736673       NA <dbl [1]>     2017-02-03T21:29:20.000Z
#> # ... with 24 more variables: `_source.parentURI` <chr>,
#> #   `_source.minElevation` <dbl>, `_source.sf-authortitle` <chr>,
#> #   `_source.xml-sitemap` <chr>, `_source.parentIdDataSet` <int>,
#> #   `_source.techKeyword` <list>, `_source.geocodes` <list>,
#> #   `_source.sp-loginOption` <int>, `_source.westBoundLongitude` <dbl>,
#> #   `_source.eastBoundLongitude` <dbl>, `_source.URI` <chr>,
#> #   `_source.northBoundLatitude` <dbl>, `_source.sp-dataStatus` <int>,
#> #   `_source.xml-thumb` <chr>, `_source.sp-hidden` <int>, `_source.xml` <chr>,
#> #   `_source.internal-source` <chr>, `_source.southBoundLatitude` <dbl>,
#> #   `_source.idDataSet` <int>, `_source.boost` <dbl>, `_source.oaiSet` <chr>,
#> #   `_source.maxElevation` <dbl>, `_source.meanPosition.lat` <dbl>,
#> #   `_source.meanPosition.lon` <dbl>

pg_search_es(query = 'water', sort = "minElevation:desc")

#> total hits: 141207

#> # A tibble: 10 × 30
#>      `_index` `_type`  `_id` `_score`      sort `_source.internal-datestamp`
#> *       <chr>   <chr>  <chr>    <lgl>    <list>                        <chr>
#> 1  pangaea_v2   panmd 316838       NA <dbl [1]>     2016-08-01T14:46:57.000Z
#> 2  pangaea_v2   panmd 316844       NA <dbl [1]>     2016-08-01T14:46:57.000Z
#> 3  pangaea_v2   panmd 316421       NA <dbl [1]>     2016-08-01T14:45:43.000Z
#> 4  pangaea_v2   panmd 316841       NA <dbl [1]>     2016-08-01T14:46:57.000Z
#> 5  pangaea_v2   panmd 316443       NA <dbl [1]>     2016-08-01T14:45:52.000Z
#> 6  pangaea_v2   panmd 316835       NA <dbl [1]>     2016-08-01T14:46:56.000Z
#> 7  pangaea_v2   panmd 316451       NA <dbl [1]>     2016-08-01T14:45:53.000Z
#> 8  pangaea_v2   panmd 317475       NA <dbl [1]>     2016-08-01T14:47:55.000Z
#> 9  pangaea_v2   panmd 316434       NA <dbl [1]>     2016-08-01T14:45:45.000Z
#> 10 pangaea_v2   panmd 316342       NA <dbl [1]>     2016-08-01T14:45:30.000Z
#> # ... with 24 more variables: `_source.maxDateTime` <chr>,
#> #   `_source.minElevation` <dbl>, `_source.sf-authortitle` <chr>,
#> #   `_source.xml-sitemap` <chr>, `_source.techKeyword` <list>,
#> #   `_source.geocodes` <list>, `_source.sp-loginOption` <int>,
#> #   `_source.westBoundLongitude` <dbl>, `_source.eastBoundLongitude` <dbl>,
#> #   `_source.URI` <chr>, `_source.northBoundLatitude` <dbl>,
#> #   `_source.sp-dataStatus` <int>, `_source.minDateTime` <chr>,
#> #   `_source.xml-thumb` <chr>, `_source.sp-hidden` <int>, `_source.xml` <chr>,
#> #   `_source.internal-source` <chr>, `_source.southBoundLatitude` <dbl>,
#> #   `_source.idDataSet` <int>, `_source.boost` <dbl>,
#> #   `_source.maxElevation` <dbl>, `_source.elevationGeocode` <chr>,
#> #   `_source.meanPosition.lat` <dbl>, `_source.meanPosition.lon` <dbl>